@@ -48,7 +48,7 @@ from .errors import ClientException, InvalidArgument
 from .file import File, VoiceMessage
 from .flags import ChannelFlags, MessageFlags
 from .invite import Invite
-from .iterators import HistoryIterator
+from .iterators import HistoryIterator, MessagePinIterator
 from .mentions import AllowedMentions
 from .partial_emoji import PartialEmoji, _EmojiTag
 from .permissions import PermissionOverwrite, Permissions
@@ -95,7 +95,7 @@ if TYPE_CHECKING:
     from .types.channel import GuildChannel as GuildChannelPayload
     from .types.channel import OverwriteType
     from .types.channel import PermissionOverwrite as PermissionOverwritePayload
-    from .ui.view import View
+    from .ui.view import BaseView
     from .user import ClientUser
 
     PartialMessageableChannel = Union[
@@ -235,7 +235,7 @@ class User(Snowflake, Protocol):
     name: str
     discriminator: str
     global_name: str | None
-    avatar: Asset
+    avatar: Asset | None
     bot: bool
 
     @property
@@ -963,7 +963,7 @@ class GuildChannel:
         if overwrite is None:
             await http.delete_channel_permissions(self.id, target.id, reason=reason)
         elif isinstance(overwrite, PermissionOverwrite):
-            (allow, deny) = overwrite.pair()
+            allow, deny = overwrite.pair()
             await http.edit_channel_permissions(
                 self.id, target.id, allow.value, deny.value, perm_type, reason=reason
             )
@@ -1355,7 +1355,7 @@ class Messageable:
         allowed_mentions: AllowedMentions = ...,
         reference: Message | MessageReference | PartialMessage = ...,
         mention_author: bool = ...,
-        view: View = ...,
+        view: BaseView = ...,
         poll: Poll = ...,
         suppress: bool = ...,
         silent: bool = ...,
@@ -1376,7 +1376,7 @@ class Messageable:
         allowed_mentions: AllowedMentions = ...,
         reference: Message | MessageReference | PartialMessage = ...,
         mention_author: bool = ...,
-        view: View = ...,
+        view: BaseView = ...,
         poll: Poll = ...,
         suppress: bool = ...,
         silent: bool = ...,
@@ -1397,7 +1397,7 @@ class Messageable:
         allowed_mentions: AllowedMentions = ...,
         reference: Message | MessageReference | PartialMessage = ...,
         mention_author: bool = ...,
-        view: View = ...,
+        view: BaseView = ...,
         poll: Poll = ...,
         suppress: bool = ...,
         silent: bool = ...,
@@ -1418,7 +1418,7 @@ class Messageable:
         allowed_mentions: AllowedMentions = ...,
         reference: Message | MessageReference | PartialMessage = ...,
         mention_author: bool = ...,
-        view: View = ...,
+        view: BaseView = ...,
         poll: Poll = ...,
         suppress: bool = ...,
         silent: bool = ...,
@@ -1497,9 +1497,9 @@ class Messageable:
             .. versionadded:: 1.4
 
         reference: Union[:class:`~discord.Message`, :class:`~discord.MessageReference`, :class:`~discord.PartialMessage`]
-            A reference to the :class:`~discord.Message` to which you are replying, this can be created using
-            :meth:`~discord.Message.to_reference` or passed directly as a :class:`~discord.Message`. You can control
-            whether this mentions the author of the referenced message using the
+            A reference to the :class:`~discord.Message` being replied to or forwarded. This can be created using
+            :meth:`~discord.Message.to_reference`.
+            When replying, you can control whether this mentions the author of the referenced message using the
             :attr:`~discord.AllowedMentions.replied_user` attribute of ``allowed_mentions`` or by
             setting ``mention_author``.
 
@@ -1509,7 +1509,7 @@ class Messageable:
             If set, overrides the :attr:`~discord.AllowedMentions.replied_user` attribute of ``allowed_mentions``.
 
             .. versionadded:: 1.6
-        view: :class:`discord.ui.View`
+        view: :class:`discord.ui.BaseView`
             A Discord UI View to add to the message.
         embeds: List[:class:`~discord.Embed`]
             A list of embeds to upload. Must be a maximum of 10.
@@ -1589,9 +1589,19 @@ class Messageable:
             allowed_mentions = allowed_mentions or AllowedMentions().to_dict()
             allowed_mentions["replied_user"] = bool(mention_author)
 
+        _reference = None
         if reference is not None:
             try:
-                reference = reference.to_message_reference_dict()
+                _reference = reference.to_message_reference_dict()
+                from .message import MessageReference
+
+                if not isinstance(reference, MessageReference):
+                    utils.warn_deprecated(
+                        f"Passing {type(reference).__name__} to reference",
+                        "MessageReference",
+                        "2.7",
+                        "3.0",
+                    )
             except AttributeError:
                 raise InvalidArgument(
                     "reference parameter must be Message, MessageReference, or"
@@ -1601,10 +1611,16 @@ class Messageable:
         if view:
             if not hasattr(view, "__discord_ui_view__"):
                 raise InvalidArgument(
-                    f"view parameter must be View not {view.__class__!r}"
+                    f"view parameter must be BaseView not {view.__class__!r}"
                 )
 
             components = view.to_components()
+            if view.is_components_v2():
+                if embeds or content:
+                    raise TypeError(
+                        "cannot send embeds or content with a view using v2 component logic"
+                    )
+                flags.is_components_v2 = True
         else:
             components = None
 
@@ -1641,7 +1657,7 @@ class Messageable:
                     nonce=nonce,
                     enforce_nonce=enforce_nonce,
                     allowed_mentions=allowed_mentions,
-                    message_reference=reference,
+                    message_reference=_reference,
                     stickers=stickers,
                     components=components,
                     flags=flags.value,
@@ -1660,7 +1676,7 @@ class Messageable:
                 nonce=nonce,
                 enforce_nonce=enforce_nonce,
                 allowed_mentions=allowed_mentions,
-                message_reference=reference,
+                message_reference=_reference,
                 stickers=stickers,
                 components=components,
                 flags=flags.value,
@@ -1669,8 +1685,10 @@ class Messageable:
 
         ret = state.create_message(channel=channel, data=data)
         if view:
-            state.store_view(view, ret.id)
+            if view.is_dispatchable():
+                state.store_view(view, ret.id)
             view.message = ret
+            view.refresh(ret.components)
 
         if delete_after is not None:
             await ret.delete(delay=delete_after)
@@ -1736,32 +1754,64 @@ class Messageable:
         data = await self._state.http.get_message(channel.id, id)
         return self._state.create_message(channel=channel, data=data)
 
-    async def pins(self) -> list[Message]:
-        """|coro|
+    def pins(
+        self,
+        *,
+        limit: int | None = 50,
+        before: SnowflakeTime | None = None,
+    ) -> MessagePinIterator:
+        """Returns a :class:`~discord.MessagePinIterator` that enables receiving the destination's pinned messages.
 
-        Retrieves all messages that are currently pinned in the channel.
+        You must have :attr:`~discord.Permissions.read_message_history` permissions to use this.
 
-        .. note::
+        .. warning::
 
-            Due to a limitation with the Discord API, the :class:`.Message`
-            objects returned by this method do not contain complete
-            :attr:`.Message.reactions` data.
+            Starting from version 3.0, `await channel.pins()` will no longer return a list of :class:`Message`. See examples below for new usage instead.
 
-        Returns
-        -------
-        List[:class:`~discord.Message`]
-            The messages that are currently pinned.
+        Parameters
+        ----------
+        limit: Optional[:class:`int`]
+            The number of pinned messages to retrieve.
+            If ``None``, retrieves every pinned message in the channel.
+        before: Optional[Union[:class:`~discord.abc.Snowflake`, :class:`datetime.datetime`]]
+            Retrieve messages pinned before this datetime.
+            If a datetime is provided, it is recommended to use a UTC aware datetime.
+            If the datetime is naive, it is assumed to be local time.
+
+        Yields
+        ------
+        :class:`~discord.MessagePin`
+            The pinned message.
 
         Raises
         ------
+        ~discord.Forbidden
+            You do not have permissions to get pinned messages.
         ~discord.HTTPException
-            Retrieving the pinned messages failed.
-        """
+            The request to get pinned messages failed.
 
-        channel = await self._get_channel()
-        state = self._state
-        data = await state.http.pins_from(channel.id)
-        return [state.create_message(channel=channel, data=m) for m in data]
+        Examples
+        --------
+
+        Usage ::
+
+            counter = 0
+            async for pin in channel.pins(limit=250):
+                if pin.message.author == client.user:
+                    counter += 1
+
+        Flattening into a list: ::
+
+            pins = await channel.pins(limit=None).flatten()
+            # pins is now a list of MessagePin...
+
+        All parameters are optional.
+        """
+        return MessagePinIterator(
+            self,
+            limit=limit,
+            before=before,
+        )
 
     def can_send(self, *objects) -> bool:
         """Returns a :class:`bool` indicating whether you have the permissions to send the object(s).
