@@ -39,6 +39,8 @@ __all__ = ("Asset",)
 if TYPE_CHECKING:
     ValidStaticFormatTypes = Literal["webp", "jpeg", "jpg", "png"]
     ValidAssetFormatTypes = Literal["webp", "jpeg", "jpg", "png", "gif"]
+    from .state import ConnectionState
+
 
 VALID_STATIC_FORMATS = frozenset({"jpeg", "jpg", "webp", "png"})
 VALID_ASSET_FORMATS = VALID_STATIC_FORMATS | {"gif"}
@@ -144,7 +146,9 @@ class Asset(AssetMixin):
 
         .. describe:: hash(x)
 
-            Returns the hash of the asset.
+            Returns the asset's url's hash.
+
+            This is equivalent to hash(:attr:`url`).
     """
 
     __slots__: tuple[str, ...] = (
@@ -200,6 +204,33 @@ class Asset(AssetMixin):
         )
 
     @classmethod
+    def _from_user_primary_guild_tag(
+        cls, state: ConnectionState, identity_guild_id: int, badge_id: str
+    ) -> Asset:
+        """Creates an Asset for a user's primary guild (tag) badge.
+
+        Parameters
+        ----------
+        state: ConnectionState
+            The connection state.
+        identity_guild_id: int
+            The ID of the guild.
+        badge_id: str
+            The badge hash/id.
+
+        Returns
+        -------
+        :class:`Asset`
+            The primary guild badge asset.
+        """
+        return cls(
+            state,
+            url=f"{Asset.BASE}/guild-tag-badges/{identity_guild_id}/{badge_id}.png?size=256",
+            key=badge_id,
+            animated=False,
+        )
+
+    @classmethod
     def _from_guild_avatar(
         cls, state, guild_id: int, member_id: int, avatar: str
     ) -> Asset:
@@ -241,6 +272,18 @@ class Asset(AssetMixin):
             url=f"{cls.BASE}/app-assets/{object_id}/store/{cover_image_hash}.png?size=1024",
             key=cover_image_hash,
             animated=False,
+        )
+
+    @classmethod
+    def _from_collectible(
+        cls, state: ConnectionState, asset: str, animated: bool = False
+    ) -> Asset:
+        name = "static.png" if not animated else "asset.webm"
+        return cls(
+            state,
+            url=f"{cls.BASE}/assets/collectibles/{asset}{name}",
+            key=asset,
+            animated=animated,
         )
 
     @classmethod
@@ -298,6 +341,14 @@ class Asset(AssetMixin):
             url=f"{cls.BASE}/guild-events/{event_id}/{cover_hash}.png",
             key=cover_hash,
             animated=False,
+        )
+
+    @classmethod
+    def _from_soundboard_sound(cls, state, sound_id: int) -> Asset:
+        return cls(
+            state,
+            url=f"{cls.BASE}/soundboard-sounds/{sound_id}",
+            key=str(sound_id),
         )
 
     def __str__(self) -> str:
