@@ -25,6 +25,8 @@ DEALINGS IN THE SOFTWARE.
 
 from __future__ import annotations
 
+import datetime
+from functools import cached_property
 from typing import TYPE_CHECKING, Any, Callable, ClassVar, Generator, TypeVar
 
 from . import enums, utils
@@ -44,8 +46,6 @@ __all__ = (
 
 
 if TYPE_CHECKING:
-    import datetime
-
     from . import abc
     from .emoji import GuildEmoji
     from .guild import Guild
@@ -209,6 +209,14 @@ def _transform_trigger_metadata(
         return AutoModTriggerMetadata.from_dict(data)
 
 
+def _transform_communication_disabled_until(
+    entry: AuditLogEntry, data: str
+) -> datetime.datetime | None:
+    if data:
+        return datetime.datetime.fromisoformat(data)
+    return None
+
+
 class AuditLogDiff:
     def __len__(self) -> int:
         return len(self.__dict__)
@@ -281,6 +289,7 @@ class AuditLogChanges:
         "trigger_metadata": (None, _transform_trigger_metadata),
         "exempt_roles": (None, _transform_roles),
         "exempt_channels": (None, _transform_channels),
+        "communication_disabled_until": (None, _transform_communication_disabled_until),
     }
 
     def __init__(
@@ -602,12 +611,12 @@ class AuditLogEntry(Hashable):
     def __repr__(self) -> str:
         return f"<AuditLogEntry id={self.id} action={self.action} user={self.user!r}>"
 
-    @utils.cached_property
+    @property
     def created_at(self) -> datetime.datetime:
         """Returns the entry's creation time in UTC."""
         return utils.snowflake_time(self.id)
 
-    @utils.cached_property
+    @property
     def target(
         self,
     ) -> (
@@ -631,24 +640,24 @@ class AuditLogEntry(Hashable):
         else:
             return converter(self._target_id)
 
-    @utils.cached_property
+    @property
     def category(self) -> enums.AuditLogActionCategory:
         """The category of the action, if applicable."""
         return self.action.category
 
-    @utils.cached_property
+    @cached_property
     def changes(self) -> AuditLogChanges:
         """The list of changes this entry has."""
         obj = AuditLogChanges(self, self._changes, state=self._state)
         del self._changes
         return obj
 
-    @utils.cached_property
+    @property
     def before(self) -> AuditLogDiff:
         """The target's prior state."""
         return self.changes.before
 
-    @utils.cached_property
+    @property
     def after(self) -> AuditLogDiff:
         """The target's subsequent state."""
         return self.changes.after
