@@ -26,7 +26,17 @@ DEALINGS IN THE SOFTWARE.
 from __future__ import annotations
 
 import datetime
-from typing import TYPE_CHECKING, Any, Callable, Iterable, Mapping, TypeVar, overload
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Callable,
+    Iterable,
+    Mapping,
+    NamedTuple,
+    Sequence,
+    TypeVar,
+    overload,
+)
 
 import discord.abc
 
@@ -39,19 +49,24 @@ from .enums import (
     InviteTarget,
     SortOrder,
     StagePrivacyLevel,
+)
+from .enums import ThreadArchiveDuration as ThreadArchiveDurationEnum
+from .enums import (
     VideoQualityMode,
+    VoiceChannelEffectAnimationType,
     VoiceRegion,
     try_enum,
 )
 from .errors import ClientException, InvalidArgument
 from .file import File
-from .flags import ChannelFlags
+from .flags import ChannelFlags, MessageFlags
 from .invite import Invite
 from .iterators import ArchivedThreadIterator
 from .mixins import Hashable
 from .object import Object
 from .partial_emoji import PartialEmoji, _EmojiTag
 from .permissions import PermissionOverwrite, Permissions
+from .soundboard import PartialSoundboardSound, SoundboardSound
 from .stage_instance import StageInstance
 from .threads import Thread
 from .utils import MISSING
@@ -67,16 +82,20 @@ __all__ = (
     "ForumChannel",
     "MediaChannel",
     "ForumTag",
+    "VoiceChannelEffectSendEvent",
 )
 
 if TYPE_CHECKING:
     from .abc import Snowflake, SnowflakeTime
+    from .embeds import Embed
     from .guild import Guild
     from .guild import GuildChannel as GuildChannelType
     from .member import Member, VoiceState
+    from .mentions import AllowedMentions
     from .message import EmojiInputType, Message, PartialMessage
     from .role import Role
     from .state import ConnectionState
+    from .sticker import GuildSticker, StickerItem
     from .types.channel import CategoryChannel as CategoryChannelPayload
     from .types.channel import DMChannel as DMChannelPayload
     from .types.channel import ForumChannel as ForumChannelPayload
@@ -85,8 +104,10 @@ if TYPE_CHECKING:
     from .types.channel import StageChannel as StageChannelPayload
     from .types.channel import TextChannel as TextChannelPayload
     from .types.channel import VoiceChannel as VoiceChannelPayload
+    from .types.channel import VoiceChannelEffectSendEvent as VoiceChannelEffectSend
     from .types.snowflake import SnowflakeList
     from .types.threads import ThreadArchiveDuration
+    from .ui.view import BaseView
     from .user import BaseUser, ClientUser, User
     from .webhook import Webhook
 
@@ -216,7 +237,7 @@ class _TextChannel(discord.abc.GuildChannel, Hashable):
 
     @property
     def _repr_attrs(self) -> tuple[str, ...]:
-        return "id", "name", "position", "nsfw", "category_id"
+        return "id", "name", "position", "category_id"
 
     def __repr__(self) -> str:
         attrs = [(val, getattr(self, val)) for val in self._repr_attrs]
@@ -812,7 +833,7 @@ class TextChannel(discord.abc.Messageable, _TextChannel):
         position: :class:`int`
             The new channel's position.
         nsfw: :class:`bool`
-            To mark the channel as NSFW or not.
+            Whether the channel is marked as NSFW.
         sync_permissions: :class:`bool`
             Whether to sync permissions with the channel's new or pre-existing
             category. Defaults to ``False``.
@@ -1039,6 +1060,8 @@ class ForumChannel(_TextChannel):
         if self.default_sort_order is not None:
             self.default_sort_order = try_enum(SortOrder, self.default_sort_order)
 
+        self.default_reaction_emoji = None
+
         reaction_emoji_ctx: dict = data.get("default_reaction_emoji")
         if reaction_emoji_ctx is not None:
             emoji_name = reaction_emoji_ctx.get("emoji_name")
@@ -1084,7 +1107,9 @@ class ForumChannel(_TextChannel):
         sync_permissions: bool = ...,
         category: CategoryChannel | None = ...,
         slowmode_delay: int = ...,
-        default_auto_archive_duration: ThreadArchiveDuration = ...,
+        default_auto_archive_duration: (
+            ThreadArchiveDuration | ThreadArchiveDurationEnum
+        ) = ...,
         default_thread_slowmode_delay: int = ...,
         default_sort_order: SortOrder = ...,
         default_reaction_emoji: GuildEmoji | int | str | None = ...,
@@ -1113,7 +1138,7 @@ class ForumChannel(_TextChannel):
         position: :class:`int`
             The new channel's position.
         nsfw: :class:`bool`
-            To mark the channel as NSFW or not.
+            Whether the channel is marked as NSFW.
         sync_permissions: :class:`bool`
             Whether to sync permissions with the channel's new or pre-existing
             category. Defaults to ``False``.
@@ -1130,6 +1155,7 @@ class ForumChannel(_TextChannel):
         default_auto_archive_duration: :class:`int`
             The new default auto archive duration in minutes for threads created in this channel.
             Must be one of ``60``, ``1440``, ``4320``, or ``10080``.
+            :class:`ThreadArchiveDuration` can be used alternatively.
         default_thread_slowmode_delay: :class:`int`
             The new default slowmode delay in seconds for threads created in this channel.
 
@@ -1181,18 +1207,20 @@ class ForumChannel(_TextChannel):
     async def create_thread(
         self,
         name: str,
-        content=None,
+        content: str | None = None,
         *,
-        embed=None,
-        embeds=None,
-        file=None,
-        files=None,
-        stickers=None,
-        delete_message_after=None,
-        nonce=None,
-        allowed_mentions=None,
-        view=None,
-        applied_tags=None,
+        embed: Embed | None = None,
+        embeds: list[Embed] | None = None,
+        file: File | None = None,
+        files: list[File] | None = None,
+        stickers: Sequence[GuildSticker | StickerItem] | None = None,
+        delete_message_after: float | None = None,
+        nonce: int | str | None = None,
+        allowed_mentions: AllowedMentions | None = None,
+        view: BaseView | None = None,
+        applied_tags: list[ForumTag] | None = None,
+        suppress: bool = False,
+        silent: bool = False,
         auto_archive_duration: ThreadArchiveDuration = MISSING,
         slowmode_delay: int = MISSING,
         reason: str | None = None,
@@ -1234,7 +1262,7 @@ class ForumChannel(_TextChannel):
             to the object, otherwise it uses the attributes set in :attr:`~discord.Client.allowed_mentions`.
             If no object is passed at all then the defaults given by :attr:`~discord.Client.allowed_mentions`
             are used instead.
-        view: :class:`discord.ui.View`
+        view: :class:`discord.ui.BaseView`
             A Discord UI View to add to the message.
         applied_tags: List[:class:`discord.ForumTag`]
             A list of tags to apply to the new thread.
@@ -1292,13 +1320,24 @@ class ForumChannel(_TextChannel):
         else:
             allowed_mentions = allowed_mentions.to_dict()
 
+        flags = MessageFlags(
+            suppress_embeds=bool(suppress),
+            suppress_notifications=bool(silent),
+        )
+
         if view:
             if not hasattr(view, "__discord_ui_view__"):
                 raise InvalidArgument(
-                    f"view parameter must be View not {view.__class__!r}"
+                    f"view parameter must be BaseView not {view.__class__!r}"
                 )
 
             components = view.to_components()
+            if view.is_components_v2():
+                if embeds or content:
+                    raise TypeError(
+                        "cannot send embeds or content with a view using v2 component logic"
+                    )
+                flags.is_components_v2 = True
         else:
             components = None
 
@@ -1337,6 +1376,7 @@ class ForumChannel(_TextChannel):
                 or self.default_auto_archive_duration,
                 rate_limit_per_user=slowmode_delay or self.slowmode_delay,
                 applied_tags=applied_tags,
+                flags=flags.value,
                 reason=reason,
             )
         finally:
@@ -1346,7 +1386,7 @@ class ForumChannel(_TextChannel):
 
         ret = Thread(guild=self.guild, state=self._state, data=data)
         msg = ret.get_partial_message(int(data["last_message_id"]))
-        if view:
+        if view and view.is_dispatchable():
             state.store_view(view, msg.id)
 
         if delete_message_after is not None:
@@ -1476,7 +1516,7 @@ class MediaChannel(ForumChannel):
         position: :class:`int`
             The new channel's position.
         nsfw: :class:`bool`
-            To mark the channel as NSFW or not.
+            Whether the channel is marked as NSFW.
         sync_permissions: :class:`bool`
             Whether to sync permissions with the channel's new or pre-existing
             category. Defaults to ``False``.
@@ -1724,6 +1764,11 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
         Extra features of the channel.
 
         .. versionadded:: 2.0
+
+    nsfw: :class:`bool`
+        Whether the channel is marked as NSFW.
+
+        .. versionadded:: 2.7
     """
 
     def __init__(
@@ -2042,6 +2087,7 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
         rtc_region: VoiceRegion | None = ...,
         video_quality_mode: VideoQualityMode = ...,
         slowmode_delay: int = ...,
+        nsfw: bool = ...,
         reason: str | None = ...,
     ) -> VoiceChannel | None: ...
 
@@ -2091,6 +2137,15 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
             The camera video quality for the voice channel's participants.
 
             .. versionadded:: 2.0
+
+        slowmode_delay: :class:`int`
+            Specifies the slowmode rate limit for user in this channel, in seconds.
+            A value of `0` disables slowmode. The maximum value possible is `21600`.
+
+        nsfw: :class:`bool`
+            Whether the channel is marked as NSFW.
+
+            .. versionadded:: 2.7
 
         Returns
         -------
@@ -2193,6 +2248,25 @@ class VoiceChannel(discord.abc.Messageable, VocalGuildChannel):
         """
         await self._state.http.set_voice_channel_status(self.id, status, reason=reason)
 
+    async def send_soundboard_sound(self, sound: PartialSoundboardSound) -> None:
+        """|coro|
+
+        Sends a soundboard sound to the voice channel.
+
+        Parameters
+        ----------
+        sound: :class:`PartialSoundboardSound`
+            The soundboard sound to send.
+
+        Raises
+        ------
+        Forbidden
+            You do not have proper permissions to send the soundboard sound.
+        HTTPException
+            Sending the soundboard sound failed.
+        """
+        await self._state.http.send_soundboard_sound(self.id, sound)
+
 
 class StageChannel(discord.abc.Messageable, VocalGuildChannel):
     """Represents a Discord guild stage channel.
@@ -2250,6 +2324,15 @@ class StageChannel(discord.abc.Messageable, VocalGuildChannel):
     last_message_id: Optional[:class:`int`]
         The ID of the last message sent to this channel. It may not always point to an existing or valid message.
         .. versionadded:: 2.5
+
+    slowmode_delay: :class:`int`
+        Specifies the slowmode rate limit for user in this channel, in seconds.
+        The maximum value possible is `21600`.
+
+    nsfw: :class:`bool`
+        Whether the channel is marked as NSFW.
+
+        .. versionadded:: 2.7
     """
 
     __slots__ = ("topic",)
@@ -2734,6 +2817,16 @@ class StageChannel(discord.abc.Messageable, VocalGuildChannel):
 
             .. versionadded:: 2.0
 
+        bitrate: :class:`int`
+            The channel's preferred audio bitrate in bits per second.
+
+        user_limit: :class:`int`
+            The channel's limit for number of members that can be in a voice channel.
+
+        slowmode_delay: :class:`int`
+            Specifies the slowmode rate limit for user in this channel, in seconds.
+            A value of `0` disables slowmode. The maximum value possible is `21600`.
+
         Returns
         -------
         Optional[:class:`.StageChannel`]
@@ -2790,12 +2883,7 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
     position: Optional[:class:`int`]
         The position in the category list. This is a number that starts at 0. e.g. the
         top category is position 0. Can be ``None`` if the channel was received in an interaction.
-    nsfw: :class:`bool`
-        If the channel is marked as "not safe for work".
 
-        .. note::
-
-            To check if the channel or the guild of that channel are marked as NSFW, consider :meth:`is_nsfw` instead.
     flags: :class:`ChannelFlags`
         Extra features of the channel.
 
@@ -2806,7 +2894,6 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
         "name",
         "id",
         "guild",
-        "nsfw",
         "_state",
         "position",
         "_overwrites",
@@ -2822,10 +2909,7 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
         self._update(guild, data)
 
     def __repr__(self) -> str:
-        return (
-            "<CategoryChannel"
-            f" id={self.id} name={self.name!r} position={self.position} nsfw={self.nsfw}>"
-        )
+        return f"<CategoryChannel id={self.id} name={self.name!r} position={self.position}>"
 
     def _update(self, guild: Guild, data: CategoryChannelPayload) -> None:
         # This data will always exist
@@ -2835,7 +2919,6 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
 
         # This data may be missing depending on how this object is being created/updated
         if not data.pop("_invoke_flag", False):
-            self.nsfw: bool = data.get("nsfw", False)
             self.position: int = data.get("position")
             self.flags: ChannelFlags = ChannelFlags._from_value(data.get("flags", 0))
             self._fill_overwrites(data)
@@ -2849,15 +2932,11 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
         """The channel's Discord type."""
         return ChannelType.category
 
-    def is_nsfw(self) -> bool:
-        """Checks if the category is NSFW."""
-        return self.nsfw
-
     @utils.copy_doc(discord.abc.GuildChannel.clone)
     async def clone(
         self, *, name: str | None = None, reason: str | None = None
     ) -> CategoryChannel:
-        return await self._clone_impl({"nsfw": self.nsfw}, name=name, reason=reason)
+        return await self._clone_impl({}, name=name, reason=reason)
 
     @overload
     async def edit(
@@ -2865,7 +2944,6 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
         *,
         name: str = ...,
         position: int = ...,
-        nsfw: bool = ...,
         overwrites: Mapping[Role | Member, PermissionOverwrite] = ...,
         reason: str | None = ...,
     ) -> CategoryChannel | None: ...
@@ -2893,8 +2971,6 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
             The new category's name.
         position: :class:`int`
             The new category's position.
-        nsfw: :class:`bool`
-            To mark the category as NSFW or not.
         reason: Optional[:class:`str`]
             The reason for editing this category. Shows up on the audit log.
         overwrites: Dict[Union[:class:`Role`, :class:`Member`, :class:`~discord.abc.Snowflake`], :class:`PermissionOverwrite`]
@@ -3041,6 +3117,25 @@ class CategoryChannel(discord.abc.GuildChannel, Hashable):
             The channel that was just created.
         """
         return await self.guild.create_forum_channel(name, category=self, **options)
+
+    @utils.deprecated(
+        since="2.7",
+        removed="3.0",
+        reference="NSFW categories are not available in the Discord API.",
+    )
+    def is_nsfw(self) -> bool:
+        return False
+
+    # TODO: Remove in 3.0
+
+    @property
+    @utils.deprecated(
+        since="2.7",
+        removed="3.0",
+        reference="NSFW categories are not available in the Discord API.",
+    )
+    def nsfw(self) -> bool:
+        return False
 
 
 DMC = TypeVar("DMC", bound="DMChannel")
@@ -3411,6 +3506,87 @@ class PartialMessageable(discord.abc.Messageable, Hashable):
         from .message import PartialMessage
 
         return PartialMessage(channel=self, id=message_id)
+
+    def __repr__(self) -> str:
+        return f"<PartialMessageable id={self.id} type={self.type!r}>"
+
+
+class VoiceChannelEffectAnimation(NamedTuple):
+    """Represents an animation that can be sent to a voice channel.
+
+    .. versionadded:: 2.7
+    """
+
+    id: int
+    type: VoiceChannelEffectAnimationType
+
+
+class VoiceChannelSoundEffect(PartialSoundboardSound): ...
+
+
+class VoiceChannelEffectSendEvent:
+    """Represents the payload for an :func:`on_voice_channel_effect_send`.
+
+    .. versionadded:: 2.7
+
+    Attributes
+    ----------
+    animation_type: :class:`int`
+        The type of animation that is being sent.
+    animation_id: :class:`int`
+        The ID of the animation that is being sent.
+    sound: Optional[:class:`SoundboardSound`]
+        The sound that is being sent, could be ``None`` if the effect is not a sound effect.
+    guild: :class:`Guild`
+        The guild in which the sound is being sent.
+    user: :class:`Member`
+        The member that sent the sound.
+    channel: :class:`VoiceChannel`
+        The voice channel in which the sound is being sent.
+    data: :class:`dict`
+        The raw data sent by the gateway.
+    """
+
+    __slots__ = (
+        "_state",
+        "animation_type",
+        "animation_id",
+        "sound",
+        "guild",
+        "user",
+        "channel",
+        "data",
+        "emoji",
+    )
+
+    def __init__(
+        self,
+        data: VoiceChannelEffectSend,
+        state: ConnectionState,
+        sound: SoundboardSound | PartialSoundboardSound | None = None,
+    ) -> None:
+        self._state = state
+        channel_id = int(data["channel_id"])
+        user_id = int(data["user_id"])
+        guild_id = int(data["guild_id"])
+        self.animation_type: VoiceChannelEffectAnimationType = try_enum(
+            VoiceChannelEffectAnimationType, data["animation_type"]
+        )
+        self.animation_id = int(data["animation_id"])
+        self.sound = sound
+        self.guild = state._get_guild(guild_id)
+        self.user = self.guild.get_member(user_id)
+        self.channel = self.guild.get_channel(channel_id)
+        self.emoji = (
+            PartialEmoji(
+                name=data["emoji"]["name"],
+                animated=data["emoji"]["animated"],
+                id=data["emoji"]["id"],
+            )
+            if data.get("emoji", None)
+            else None
+        )
+        self.data = data
 
 
 def _guild_channel_factory(channel_type: int):
