@@ -26,8 +26,25 @@ from __future__ import annotations
 
 import inspect
 import logging
+import sys
+import types
+from collections.abc import Awaitable, Callable, Iterable
 from enum import Enum
-from typing import TYPE_CHECKING, Literal, Optional, Type, Union
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Literal,
+    Optional,
+    Type,
+    TypeVar,
+    Union,
+    get_args,
+)
+
+if sys.version_info >= (3, 12):
+    from typing import TypeAliasType
+else:
+    from typing_extensions import TypeAliasType
 
 from ..abc import GuildChannel, Mentionable
 from ..channel import (
@@ -40,13 +57,14 @@ from ..channel import (
     Thread,
     VoiceChannel,
 )
-from ..commands import ApplicationContext
+from ..commands import ApplicationContext, AutocompleteContext
 from ..enums import ChannelType
 from ..enums import Enum as DiscordEnum
 from ..enums import SlashCommandOptionType
 from ..utils import MISSING, basic_autocomplete
 
 if TYPE_CHECKING:
+    from ..cog import Cog
     from ..ext.commands import Converter
     from ..member import Member
     from ..message import Attachment
@@ -71,6 +89,25 @@ if TYPE_CHECKING:
         Type[Enum],
         Type[DiscordEnum],
     ]
+
+    AutocompleteReturnType = Union[
+        Iterable["OptionChoice"], Iterable[str], Iterable[int], Iterable[float]
+    ]
+    T = TypeVar("T", bound=AutocompleteReturnType)
+    MaybeAwaitable = Union[T, Awaitable[T]]
+    AutocompleteFunction = Union[
+        Callable[[AutocompleteContext], MaybeAwaitable[AutocompleteReturnType]],
+        Callable[[Cog, AutocompleteContext], MaybeAwaitable[AutocompleteReturnType]],
+        Callable[
+            [AutocompleteContext, Any],  # pyright: ignore [reportExplicitAny]
+            MaybeAwaitable[AutocompleteReturnType],
+        ],
+        Callable[
+            [Cog, AutocompleteContext, Any],  # pyright: ignore [reportExplicitAny]
+            MaybeAwaitable[AutocompleteReturnType],
+        ],
+    ]
+
 
 __all__ = (
     "ThreadOption",
@@ -148,15 +185,6 @@ class Option:
     max_length: Optional[:class:`int`]
         The maximum length of the string that can be entered. Must be between 1 and 6000 (inclusive).
         Only applies to Options with an :attr:`input_type` of :class:`str`.
-    autocomplete: Optional[Callable[[:class:`.AutocompleteContext`], Awaitable[Union[Iterable[:class:`.OptionChoice`], Iterable[:class:`str`], Iterable[:class:`int`], Iterable[:class:`float`]]]]]
-        The autocomplete handler for the option. Accepts a callable (sync or async)
-        that takes a single argument of :class:`AutocompleteContext`.
-        The callable must return an iterable of :class:`str` or :class:`OptionChoice`.
-        Alternatively, :func:`discord.utils.basic_autocomplete` may be used in place of the callable.
-
-        .. note::
-
-            Does not validate the input value against the autocomplete results.
     channel_types: list[:class:`discord.ChannelType`] | None
         A list of channel types that can be selected in this option.
         Only applies to Options with an :attr:`input_type` of :class:`discord.SlashCommandOptionType.channel`.
@@ -196,6 +224,8 @@ class Option:
         if self.name is not None:
             self.name = str(self.name)
         self._parameter_name = self.name  # default
+        input_type = self._parse_type_alias(input_type)
+        input_type = self._strip_none_type(input_type)
         self._raw_type: InputType | tuple = input_type
 
         enum_choices = []
@@ -225,7 +255,9 @@ class Option:
         self.description = description or "No description provided"
         self.channel_types: list[ChannelType] = kwargs.pop("channel_types", [])
 
-        if isinstance(input_type, SlashCommandOptionType):
+        if self.channel_types:
+            self.input_type = SlashCommandOptionType.channel
+        elif isinstance(input_type, SlashCommandOptionType):
             self.input_type = input_type
         else:
             from ..ext.commands import Converter
@@ -274,6 +306,8 @@ class Option:
         )
         self.default = kwargs.pop("default", None)
 
+        self._autocomplete: AutocompleteFunction | None = None
+        self._autocomplete_is_instance_method: bool = False
         self.autocomplete = kwargs.pop("autocomplete", None)
         if len(enum_choices) > 25:
             self.choices: list[OptionChoice] = []
@@ -363,6 +397,39 @@ class Option:
         if input_type is None:
             raise TypeError("input_type cannot be NoneType.")
 
+    @staticmethod
+    def _parse_type_alias(input_type: InputType) -> InputType:
+        if isinstance(input_type, TypeAliasType):
+            return input_type.__value__
+        return input_type
+
+    @staticmethod
+    def _strip_none_type(input_type):
+        if isinstance(input_type, SlashCommandOptionType):
+            return input_type
+
+        if input_type is type(None):
+            raise TypeError("Option type cannot be only NoneType")
+
+        args = ()
+        if isinstance(input_type, types.UnionType):
+            args = get_args(input_type)
+        elif getattr(input_type, "__origin__", None) is Union:
+            args = get_args(input_type)
+        elif isinstance(input_type, tuple):
+            args = input_type
+
+        if args:
+            filtered = tuple(t for t in args if t is not type(None))
+            if not filtered:
+                raise TypeError("Option type cannot be only NoneType")
+            if len(filtered) == 1:
+                return filtered[0]
+
+            return filtered
+
+        return input_type
+
     def to_dict(self) -> dict:
         as_dict = {
             "name": self.name,
@@ -391,6 +458,43 @@ class Option:
 
     def __repr__(self):
         return f"<discord.commands.{self.__class__.__name__} name={self.name}>"
+
+    @property
+    def autocomplete(self) -> AutocompleteFunction | None:
+        """
+        The autocomplete handler for the option. Accepts a callable (sync or async)
+        that takes a single required argument of :class:`AutocompleteContext` or two arguments
+        of :class:`discord.Cog` (being the command's cog) and :class:`AutocompleteContext`.
+        The callable must return an iterable of :class:`str` or :class:`OptionChoice`.
+        Alternatively, :func:`discord.utils.basic_autocomplete` may be used in place of the callable.
+
+        Returns
+        -------
+        Optional[AutocompleteFunction]
+
+        .. versionchanged:: 2.7
+
+        .. note::
+            Does not validate the input value against the autocomplete results.
+        """
+        return self._autocomplete
+
+    @autocomplete.setter
+    def autocomplete(self, value: AutocompleteFunction | None) -> None:
+        self._autocomplete = value
+        # this is done here so it does not have to be computed every time the autocomplete is invoked
+        if self._autocomplete is not None:
+            self._autocomplete_is_instance_method = (
+                sum(
+                    1
+                    for param in inspect.signature(
+                        self._autocomplete
+                    ).parameters.values()
+                    if param.default == param.empty
+                    and param.kind not in (param.VAR_POSITIONAL, param.VAR_KEYWORD)
+                )
+                == 2
+            )
 
 
 class OptionChoice:
