@@ -25,7 +25,7 @@ DEALINGS IN THE SOFTWARE.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Iterator
+from typing import TYPE_CHECKING, Any, Iterator, Literal
 
 from .asset import Asset, AssetMixin
 from .partial_emoji import PartialEmoji, _EmojiTag
@@ -49,7 +49,6 @@ if TYPE_CHECKING:
 
 
 class BaseEmoji(_EmojiTag, AssetMixin):
-
     __slots__: tuple[str, ...] = (
         "require_colons",
         "animated",
@@ -107,8 +106,25 @@ class BaseEmoji(_EmojiTag, AssetMixin):
     @property
     def url(self) -> str:
         """Returns the URL of the emoji."""
-        fmt = "gif" if self.animated else "png"
-        return f"{Asset.BASE}/emojis/{self.id}.{fmt}"
+        url = f"{Asset.BASE}/emojis/{self.id}.{self.extension}"
+        if self.animated:
+            url += "?animated=true"
+        return url
+
+    @property
+    def mention(self) -> str:
+        """Return a string that allows you to mention the emoji in a message."""
+        if self.animated:
+            return f"<a:{self.name}:{self.id}>"
+        return f"<:{self.name}:{self.id}>"
+
+    @property
+    def extension(self) -> Literal["webp", "png"]:
+        """Return the file extension of the emoji.
+
+        .. versionadded:: 2.7.1
+        """
+        return "webp" if self.animated else "png"
 
 
 class GuildEmoji(BaseEmoji):
@@ -345,7 +361,7 @@ class AppEmoji(BaseEmoji):
         super().__init__(state=state, data=data)
 
     def __repr__(self) -> str:
-        return "<AppEmoji" f" id={self.id} name={self.name!r} animated={self.animated}>"
+        return f"<AppEmoji id={self.id} name={self.name!r} animated={self.animated}>"
 
     @property
     def guild(self) -> Guild:
