@@ -51,6 +51,7 @@ from ..commands import (
     Converter,
     Group,
     GuildChannelConverter,
+    MemberConverter,
     RoleConverter,
     UserConverter,
 )
@@ -98,26 +99,12 @@ class BridgeExtCommand(Command):
     def __init__(self, func, **kwargs):
         super().__init__(func, **kwargs)
 
-        # TODO: v2.7: Remove backwards support for Option in bridge commands.
-        for name, option in self.params.items():
+        for option in self.params.values():
             if isinstance(option.annotation, Option) and not isinstance(
                 option.annotation, BridgeOption
             ):
-                # Warn not to do this
-                warn_deprecated(
-                    "Using Option for bridge commands",
-                    "BridgeOption",
-                    "2.5",
-                    "2.7",
-                    reference="https://github.com/Pycord-Development/pycord/pull/2417",
-                    stacklevel=6,
-                )
-                # Override the convert method of the parameter's annotated Option.
-                # We can use the convert method from BridgeOption, and bind "self"
-                # using a manual invocation of the descriptor protocol.
-                # Definitely not a good approach, but gets the job done until removal.
-                self.params[name].annotation.convert = BridgeOption.convert.__get__(
-                    self.params[name].annotation
+                raise TypeError(
+                    f"{option.annotation.__class__.__name__} is not supported in bridge commands. Use BridgeOption instead."
                 )
 
     async def dispatch_error(self, ctx: BridgeExtContext, error: Exception) -> None:
@@ -184,6 +171,8 @@ class BridgeCommand:
         The prefix-based version of this bridge command.
     """
 
+    __bridge__: bool = True
+
     __special_attrs__ = ["slash_variant", "ext_variant", "parent"]
 
     def __init__(self, callback, **kwargs):
@@ -198,8 +187,10 @@ class BridgeCommand:
     @property
     def name_localizations(self) -> dict[str, str] | None:
         """Returns name_localizations from :attr:`slash_variant`
-        You can edit/set name_localizations directly with
+           You can edit/set name_localizations directly with
+
         .. code-block:: python3
+
             bridge_command.name_localizations["en-UK"] = ...  # or any other locale
             # or
             bridge_command.name_localizations = {"en-UK": ..., "fr-FR": ...}
@@ -213,8 +204,10 @@ class BridgeCommand:
     @property
     def description_localizations(self) -> dict[str, str] | None:
         """Returns description_localizations from :attr:`slash_variant`
-        You can edit/set description_localizations directly with
+           You can edit/set description_localizations directly with
+
         .. code-block:: python3
+
             bridge_command.description_localizations["en-UK"] = ...  # or any other locale
             # or
             bridge_command.description_localizations = {"en-UK": ..., "fr-FR": ...}
@@ -580,13 +573,21 @@ def has_permissions(**perms: bool):
 
 
 class MentionableConverter(Converter):
-    """A converter that can convert a mention to a user or a role."""
+    """A converter that can convert a mention to a member, a user or a role."""
 
     async def convert(self, ctx, argument):
         try:
             return await RoleConverter().convert(ctx, argument)
         except BadArgument:
-            return await UserConverter().convert(ctx, argument)
+            pass
+
+        if ctx.guild:
+            try:
+                return await MemberConverter().convert(ctx, argument)
+            except BadArgument:
+                pass
+
+        return await UserConverter().convert(ctx, argument)
 
 
 class AttachmentConverter(Converter):
@@ -614,6 +615,7 @@ BRIDGE_CONVERTER_MAPPING = {
     SlashCommandOptionType.mentionable: MentionableConverter,
     SlashCommandOptionType.number: float,
     SlashCommandOptionType.attachment: AttachmentConverter,
+    discord.Member: MemberConverter,
 }
 
 
@@ -622,16 +624,24 @@ class BridgeOption(Option, Converter):
     command option and a prefixed command argument for bridge commands.
     """
 
+    def __init__(self, input_type, *args, **kwargs):
+        self.converter = kwargs.pop("converter", None)
+        super().__init__(input_type, *args, **kwargs)
+
+        self.converter = self.converter or BRIDGE_CONVERTER_MAPPING.get(input_type)
+
     async def convert(self, ctx, argument: str) -> Any:
         try:
             if self.converter is not None:
-                converted = await self.converter.convert(ctx, argument)
+                converted = await self.converter().convert(ctx, argument)
             else:
-                converter = BRIDGE_CONVERTER_MAPPING[self.input_type]
-                if issubclass(converter, Converter):
+                converter = BRIDGE_CONVERTER_MAPPING.get(self.input_type)
+                if isinstance(converter, type) and issubclass(converter, Converter):
                     converted = await converter().convert(ctx, argument)  # type: ignore # protocol class
-                else:
+                elif callable(converter):
                     converted = converter(argument)
+                else:
+                    raise TypeError(f"Invalid converter: {converter}")
 
             if self.choices:
                 choices_names: list[str | int | float] = [
