@@ -333,22 +333,22 @@ class MessageFlags(BaseFlags):
     @flag_value
     def crossposted(self):
         """:class:`bool`: Returns ``True`` if the message is the original crossposted message."""
-        return 1
+        return 1 << 0
 
     @flag_value
     def is_crossposted(self):
         """:class:`bool`: Returns ``True`` if the message was crossposted from another channel."""
-        return 2
+        return 1 << 1
 
     @flag_value
     def suppress_embeds(self):
         """:class:`bool`: Returns ``True`` if the message's embeds have been suppressed."""
-        return 4
+        return 1 << 2
 
     @flag_value
     def source_message_deleted(self):
         """:class:`bool`: Returns ``True`` if the source message for this crosspost has been deleted."""
-        return 8
+        return 1 << 3
 
     @flag_value
     def urgent(self):
@@ -356,7 +356,7 @@ class MessageFlags(BaseFlags):
 
         An urgent message is one sent by Discord Trust and Safety.
         """
-        return 16
+        return 1 << 4
 
     @flag_value
     def has_thread(self):
@@ -364,7 +364,7 @@ class MessageFlags(BaseFlags):
 
         .. versionadded:: 2.0
         """
-        return 32
+        return 1 << 5
 
     @flag_value
     def ephemeral(self):
@@ -372,7 +372,7 @@ class MessageFlags(BaseFlags):
 
         .. versionadded:: 2.0
         """
-        return 64
+        return 1 << 6
 
     @flag_value
     def loading(self):
@@ -382,7 +382,7 @@ class MessageFlags(BaseFlags):
 
         .. versionadded:: 2.0
         """
-        return 128
+        return 1 << 7
 
     @flag_value
     def failed_to_mention_some_roles_in_thread(self):
@@ -390,7 +390,7 @@ class MessageFlags(BaseFlags):
 
         .. versionadded:: 2.0
         """
-        return 256
+        return 1 << 8
 
     @flag_value
     def suppress_notifications(self):
@@ -401,7 +401,7 @@ class MessageFlags(BaseFlags):
         .. versionadded:: 2.4
         """
 
-        return 4096
+        return 1 << 12
 
     @flag_value
     def is_voice_message(self):
@@ -409,7 +409,23 @@ class MessageFlags(BaseFlags):
 
         .. versionadded:: 2.5
         """
-        return 8192
+        return 1 << 13
+
+    @flag_value
+    def is_components_v2(self):
+        """:class:`bool`: Returns ``True`` if this message has v2 components. This flag disables sending `content`, `embed`, and `embeds`.
+
+        .. versionadded:: 2.7
+        """
+        return 1 << 15
+
+    @flag_value
+    def has_snapshot(self):
+        """:class:`bool`: Returns ``True`` if this message has a snapshot from message forwarding.
+
+        .. versionadded:: 2.7
+        """
+        return 1 << 14
 
 
 @fill_with_flags()
