@@ -26,6 +26,7 @@ DEALINGS IN THE SOFTWARE.
 from __future__ import annotations
 
 import copy
+import datetime
 import unicodedata
 from typing import (
     TYPE_CHECKING,
@@ -36,9 +37,12 @@ from typing import (
     Optional,
     Sequence,
     Tuple,
+    TypeVar,
     Union,
     overload,
 )
+
+from typing_extensions import override
 
 from . import abc, utils
 from .asset import Asset
@@ -56,16 +60,19 @@ from .enums import (
     EntitlementOwnerType,
     NotificationLevel,
     NSFWLevel,
+    OnboardingMode,
     ScheduledEventLocationType,
     ScheduledEventPrivacyLevel,
+    SortOrder,
     VerificationLevel,
     VideoQualityMode,
     VoiceRegion,
     try_enum,
 )
-from .errors import ClientException, InvalidArgument, InvalidData
+from .errors import ClientException, HTTPException, InvalidArgument, InvalidData
 from .file import File
 from .flags import SystemChannelFlags
+from .incidents import IncidentsData
 from .integrations import Integration, _integration_factory
 from .invite import Invite
 from .iterators import (
@@ -79,16 +86,18 @@ from .mixins import Hashable
 from .monetization import Entitlement
 from .onboarding import Onboarding
 from .permissions import PermissionOverwrite
-from .role import Role
+from .role import Role, RoleColours
 from .scheduled_events import ScheduledEvent, ScheduledEventLocation
+from .soundboard import SoundboardSound
 from .stage_instance import StageInstance
 from .sticker import GuildSticker
 from .threads import Thread, ThreadMember
 from .user import User
+from .utils import _D, _FETCHABLE
 from .welcome_screen import WelcomeScreen, WelcomeScreenChannel
 from .widget import Widget
 
-__all__ = ("Guild",)
+__all__ = ("BanEntry", "Guild", "GuildRoleCounts")
 
 MISSING = utils.MISSING
 
@@ -103,12 +112,17 @@ if TYPE_CHECKING:
         TextChannel,
         VoiceChannel,
     )
+    from .onboarding import OnboardingPrompt
     from .permissions import Permissions
     from .state import ConnectionState
     from .template import Template
     from .types.guild import Ban as BanPayload
     from .types.guild import Guild as GuildPayload
-    from .types.guild import GuildFeature, MFALevel
+    from .types.guild import (
+        GuildFeature,
+        MFALevel,
+    )
+    from .types.guild import ModifyIncidents as ModifyIncidentsPayload
     from .types.member import Member as MemberPayload
     from .types.threads import Thread as ThreadPayload
     from .types.voice import GuildVoiceState
@@ -121,6 +135,8 @@ if TYPE_CHECKING:
     ]
     ByCategoryItem = Tuple[Optional[CategoryChannel], List[GuildChannel]]
 
+T = TypeVar("T")
+
 
 class BanEntry(NamedTuple):
     reason: str | None
@@ -130,8 +146,84 @@ class BanEntry(NamedTuple):
 class _GuildLimit(NamedTuple):
     emoji: int
     stickers: int
+    soundboard: int
     bitrate: float
     filesize: int
+
+
+class GuildRoleCounts(dict[int, int]):
+    """A dictionary subclass that maps role IDs to their member counts.
+
+    This class allows accessing member counts by either role ID (:class:`int`) or by
+    a Snowflake object (which has an ``.id`` attribute).
+
+    .. versionadded:: 2.7
+    """
+
+    @override
+    def __repr__(self):
+        return f"<GuildRoleCounts {super().__repr__()}>"
+
+    @override
+    def __getitem__(self, key: int | abc.Snowflake) -> int:
+        """Get the member count for a role.
+
+        Parameters
+        ----------
+        key: Union[:class:`int`, :class:`~discord.abc.Snowflake`]
+            The role ID or a Snowflake object (e.g., a :class:`Role`).
+
+        Returns
+        -------
+        :class:`int`
+            The member count for the role.
+
+        Raises
+        ------
+        KeyError
+            The role ID was not found.
+        """
+        if isinstance(key, abc.Snowflake):
+            key = key.id
+        return super().__getitem__(key)
+
+    @override
+    def get(self, key: int | abc.Snowflake, default: T = None) -> int | T:
+        """Get the member count for a role, returning a default if not found.
+
+        Parameters
+        ----------
+        key: Union[:class:`int`, :class:`~discord.abc.Snowflake`]
+            The role ID or a Snowflake object (e.g., a :class:`Role`).
+        default: Any
+            The value to return if the role ID is not found.
+
+        Returns
+        -------
+        Optional[:class:`int`]
+            The member count for the role, or ``default`` if the role is not present.
+        """
+        if isinstance(key, abc.Snowflake):
+            key = key.id
+        return super().get(key, default)
+
+    @override
+    def __contains__(self, key: int | abc.Snowflake) -> bool:
+        """Check if a role ID or Snowflake object is in the counts.
+
+        Parameters
+        ----------
+        key: Union[:class:`int`, :class:`~discord.abc.Snowflake`]
+            The role ID or a Snowflake object (e.g., a :class:`Role`).
+
+        Returns
+        -------
+        :class:`bool`
+            ``True`` if the role ID is present, ``False`` otherwise.
+        """
+        if isinstance(key, abc.Snowflake):
+            key = key.id
+        return super().__contains__(key)
 
 
 class Guild(Hashable):
@@ -173,8 +265,6 @@ class Guild(Hashable):
         The channel that denotes the AFK channel. ``None`` if it doesn't exist.
     id: :class:`int`
         The guild's ID.
-    invites_disabled: :class:`bool`
-        Indicates if the guild invites are disabled.
     owner_id: :class:`int`
         The guild owner's ID. Use :attr:`Guild.owner` instead.
     unavailable: :class:`bool`
@@ -241,6 +331,11 @@ class Guild(Hashable):
         with ``with_counts=True``.
 
         .. versionadded:: 2.0
+
+    incidents_data: Optional[:class:`IncidentsData`]
+        The incidents data for the guild.
+
+        .. versionadded:: 2.7
     """
 
     __slots__ = (
@@ -286,14 +381,26 @@ class Guild(Hashable):
         "_threads",
         "approximate_member_count",
         "approximate_presence_count",
+        "_sounds",
+        "incidents_data",
     )
 
     _PREMIUM_GUILD_LIMITS: ClassVar[dict[int | None, _GuildLimit]] = {
-        None: _GuildLimit(emoji=50, stickers=5, bitrate=96e3, filesize=10_485_760),
-        0: _GuildLimit(emoji=50, stickers=5, bitrate=96e3, filesize=10_485_760),
-        1: _GuildLimit(emoji=100, stickers=15, bitrate=128e3, filesize=10_485_760),
-        2: _GuildLimit(emoji=150, stickers=30, bitrate=256e3, filesize=52_428_800),
-        3: _GuildLimit(emoji=250, stickers=60, bitrate=384e3, filesize=104_857_600),
+        None: _GuildLimit(
+            emoji=50, stickers=5, soundboard=8, bitrate=96e3, filesize=10_485_760
+        ),
+        0: _GuildLimit(
+            emoji=50, stickers=5, soundboard=8, bitrate=96e3, filesize=10_485_760
+        ),
+        1: _GuildLimit(
+            emoji=100, stickers=15, soundboard=24, bitrate=128e3, filesize=10_485_760
+        ),
+        2: _GuildLimit(
+            emoji=150, stickers=30, soundboard=36, bitrate=256e3, filesize=52_428_800
+        ),
+        3: _GuildLimit(
+            emoji=250, stickers=60, soundboard=48, bitrate=384e3, filesize=104_857_600
+        ),
     }
 
     def __init__(self, *, data: GuildPayload, state: ConnectionState):
@@ -308,6 +415,7 @@ class Guild(Hashable):
         self._voice_states: dict[int, VoiceState] = {}
         self._threads: dict[int, Thread] = {}
         self._state: ConnectionState = state
+        self._sounds: dict[int, SoundboardSound] = {}
         self._from_data(data)
 
     def _add_channel(self, channel: GuildChannel, /) -> None:
@@ -550,6 +658,140 @@ class Guild(Hashable):
         for obj in guild.get("voice_states", []):
             self._update_voice_state(obj, int(obj["channel_id"]))
 
+        for sound in guild.get("soundboard_sounds", []):
+            sound = SoundboardSound(state=state, http=state.http, data=sound)
+            self._add_sound(sound)
+
+        incidents_payload = guild.get("incidents_data")
+        self.incidents_data: IncidentsData | None = (
+            IncidentsData(data=incidents_payload)
+            if incidents_payload is not None
+            else None
+        )
+
+    def _add_sound(self, sound: SoundboardSound) -> None:
+        self._sounds[sound.id] = sound
+        self._state._add_sound(sound)
+
+    def _remove_sound(self, sound_id: int) -> None:
+        self._sounds.pop(sound_id, None)
+
+    async def fetch_sounds(self) -> list[SoundboardSound]:
+        """|coro|
+        Fetches all the soundboard sounds in the guild.
+
+        .. versionadded:: 2.7
+
+        Returns
+        -------
+        List[:class:`SoundboardSound`]
+            The sounds in the guild.
+        """
+        data = await self._state.http.get_all_guild_sounds(self.id)
+        return [
+            SoundboardSound(
+                state=self._state,
+                http=self._state.http,
+                data=sound,
+            )
+            for sound in data["items"]
+        ]
+
+    async def fetch_sound(self, sound_id: int) -> SoundboardSound:
+        """|coro|
+        Fetches a soundboard sound in the guild.
+
+        .. versionadded:: 2.7
+
+        Parameters
+        ----------
+        sound_id: :class:`int`
+            The ID of the sound.
+
+        Returns
+        -------
+        :class:`SoundboardSound`
+            The sound.
+        """
+        data = await self._state.http.get_guild_sound(self.id, sound_id)
+        return SoundboardSound(
+            state=self._state,
+            http=self._state.http,
+            data=data,
+        )
+
+    async def create_sound(
+        self,
+        name: str,
+        sound: bytes,
+        volume: float = 1.0,
+        emoji: PartialEmoji | GuildEmoji | str | None = None,
+        reason: str | None = None,
+    ) -> SoundboardSound:
+        """|coro|
+        Creates a :class:`SoundboardSound` in the guild.
+        You must have :attr:`Permissions.manage_expressions` permission to use this.
+
+        .. versionadded:: 2.7
+
+        Parameters
+        ----------
+        name: :class:`str`
+            The name of the sound.
+        sound: :class:`bytes`
+            The :term:`py:bytes-like object` representing the sound data.
+            Only MP3 sound files that are less than 5.2 seconds long are supported.
+        volume: :class:`float`
+            The volume of the sound. Defaults to 1.0.
+        emoji: Optional[Union[:class:`PartialEmoji`, :class:`GuildEmoji`, :class:`str`]]
+            The emoji of the sound.
+        reason: Optional[:class:`str`]
+            The reason for creating this sound. Shows up on the audit log.
+
+        Returns
+        -------
+        :class:`SoundboardSound`
+            The created sound.
+
+        Raises
+        ------
+        :exc:`HTTPException`
+            Creating the sound failed.
+        :exc:`Forbidden`
+            You do not have permissions to create sounds.
+        """
+
+        payload: dict[str, Any] = {
+            "name": name,
+            "sound": utils._bytes_to_base64_data(sound),
+            "volume": volume,
+            "emoji_id": None,
+            "emoji_name": None,
+        }
+
+        if emoji is not None:
+            if isinstance(emoji, _EmojiTag):
+                partial_emoji = emoji._to_partial()
+            elif isinstance(emoji, str):
+                partial_emoji = PartialEmoji.from_str(emoji)
+            else:
+                partial_emoji = None
+
+            if partial_emoji is not None:
+                if partial_emoji.id is None:
+                    payload["emoji_name"] = partial_emoji.name
+                else:
+                    payload["emoji_id"] = partial_emoji.id
+
+        data = await self._state.http.create_guild_sound(
+            self.id, reason=reason, **payload
+        )
+        return SoundboardSound(
+            state=self._state,
+            http=self._state.http,
+            data=data,
+        )
+
     # TODO: refactor/remove?
     def _sync(self, data: GuildPayload) -> None:
         try:
@@ -674,6 +916,17 @@ class Guild(Hashable):
         """
         r = [ch for ch in self._channels.values() if isinstance(ch, CategoryChannel)]
         r.sort(key=lambda c: (c.position or -1, c.id))
+        return r
+
+    @property
+    def sounds(self) -> list[SoundboardSound]:
+        """A list of soundboard sounds that belong to this guild.
+
+        .. versionadded:: 2.7
+
+        This is sorted by the position and are in UI order from top to bottom.
+        """
+        r = list(self._sounds.values())
         return r
 
     def by_category(self) -> list[ByCategoryItem]:
@@ -827,6 +1080,17 @@ class Guild(Hashable):
         )
 
     @property
+    def soundboard_limit(self) -> int:
+        """The maximum number of soundboard slots this guild has.
+
+        .. versionadded:: 2.7
+        """
+        more_soundboard = 48 if "MORE_SOUNDBOARD" in self.features else 0
+        return max(
+            more_soundboard, self._PREMIUM_GUILD_LIMITS[self.premium_tier].soundboard
+        )
+
+    @property
     def bitrate_limit(self) -> int:
         """The maximum bitrate for voice channels this guild can have."""
         vip_guild = (
@@ -863,6 +1127,49 @@ class Guild(Hashable):
         """
         return self._members.get(user_id)
 
+    async def get_or_fetch(
+        self: Guild,
+        object_type: type[_FETCHABLE],
+        object_id: int | None,
+        default: _D = None,
+    ) -> _FETCHABLE | _D | None:
+        """
+        Shortcut method to get data from this guild either by returning the cached version,
+        or if it does not exist, attempting to fetch it from the API.
+
+        Parameters
+        ----------
+        object_type: Type[:class:`VoiceChannel` | :class:`TextChannel` | :class:`ForumChannel` | :class:`StageChannel` | :class:`CategoryChannel` | :class:`Thread` | :class:`Role` | :class:`Member` | :class:`GuildEmoji`]
+            Type of object to fetch or get.
+
+        object_id: :class:`int` | :data:`None`
+            ID of the object to get. If :data:`None`, returns `default` if provided, otherwise :data:`None`.
+
+        default: Any | :data:`None`
+            The value to return instead of raising if fetching fails or if `object_id` is :data:`None`.
+
+        Returns
+        -------
+        :class:`VoiceChannel` | :class:`TextChannel` | :class:`ForumChannel` | :class:`StageChannel` | :class:`CategoryChannel` | :class:`Thread` | :class:`Role` | :class:`Member` | :class:`GuildEmoji` | :data:`None`
+            The object if found, or `default` if provided when not found.
+
+        Raises
+        ------
+        :exc:`TypeError`
+            Raised when required parameters are missing or invalid types are provided.
+        :exc:`InvalidArgument`
+            Raised when an unsupported or incompatible object type is used.
+        """
+        try:
+            return await utils.get_or_fetch(
+                obj=self,
+                object_type=object_type,
+                object_id=object_id,
+                default=default,
+            )
+        except (HTTPException, ValueError, InvalidData):
+            return default
+
     @property
     def premium_subscribers(self) -> list[Member]:
         """A list of members who have "boosted" this guild."""
@@ -891,6 +1198,44 @@ class Guild(Hashable):
             The role or ``None`` if not found.
         """
         return self._roles.get(role_id)
+
+    async def fetch_roles_member_counts(self) -> GuildRoleCounts:
+        """|coro|
+        Fetches a mapping of role IDs to their member counts for this guild.
+
+        .. versionadded:: 2.7
+
+        Returns
+        -------
+        :class:`GuildRoleCounts`
+            A mapping of role IDs to their member counts. Can be accessed
+            with either role IDs (:class:`int`) or Snowflake objects (e.g., :class:`Role`).
+
+        Raises
+        ------
+        :exc:`HTTPException`
+            Fetching the role member counts failed.
+
+        Examples
+        --------
+
+        Getting member counts using role IDs:
+
+        .. code-block:: python3
+
+            counts = await guild.fetch_roles_member_counts()
+            member_count = counts[123456789]
+
+        Using a role object:
+
+        .. code-block:: python3
+
+            counts = await guild.fetch_roles_member_counts()
+            role = guild.get_role(123456789)
+            member_count = counts[role]
+        """
+        r = await self._state.http.get_roles_member_counts(self.id)
+        return GuildRoleCounts({int(role_id): count for role_id, count in r.items()})
 
     @property
     def default_role(self) -> Role:
@@ -1027,7 +1372,7 @@ class Guild(Hashable):
 
     @property
     def invites_disabled(self) -> bool:
-        """Returns a boolean indicating if the guild invites are disabled."""
+        """A boolean indicating whether the guild invites are disabled."""
         return "INVITES_DISABLED" in self.features
 
     def get_member_named(self, name: str, /) -> Member | None:
@@ -1130,6 +1475,8 @@ class Guild(Hashable):
         slowmode_delay: int = MISSING,
         nsfw: bool = MISSING,
         overwrites: dict[Role | Member, PermissionOverwrite] = MISSING,
+        default_thread_slowmode_delay: int | None = MISSING,
+        default_auto_archive_duration: int = MISSING,
     ) -> TextChannel:
         """|coro|
 
@@ -1166,11 +1513,21 @@ class Guild(Hashable):
             The new channel's topic.
         slowmode_delay: :class:`int`
             Specifies the slowmode rate limit for user in this channel, in seconds.
-            The maximum value possible is `21600`.
+            A value of `0` disables slowmode. The maximum value possible is `21600`.
         nsfw: :class:`bool`
-            To mark the channel as NSFW or not.
+            Whether the channel is marked as NSFW.
         reason: Optional[:class:`str`]
             The reason for creating this channel. Shows up on the audit log.
+
+        default_thread_slowmode_delay: Optional[:class:`int`]
+            The initial slowmode delay to set on newly created threads in this channel.
+
+            .. versionadded:: 2.7
+
+        default_auto_archive_duration: :class:`int`
+            The default auto archive duration in minutes for threads created in this channel.
+
+            .. versionadded:: 2.7
 
         Returns
         -------
@@ -1220,6 +1577,12 @@ class Guild(Hashable):
         if nsfw is not MISSING:
             options["nsfw"] = nsfw
 
+        if default_thread_slowmode_delay is not MISSING:
+            options["default_thread_slowmode_delay"] = default_thread_slowmode_delay
+
+        if default_auto_archive_duration is not MISSING:
+            options["default_auto_archive_duration"] = default_auto_archive_duration
+
         data = await self._create_channel(
             name,
             overwrites=overwrites,
@@ -1246,6 +1609,8 @@ class Guild(Hashable):
         rtc_region: VoiceRegion | None = MISSING,
         video_quality_mode: VideoQualityMode = MISSING,
         overwrites: dict[Role | Member, PermissionOverwrite] = MISSING,
+        slowmode_delay: int = MISSING,
+        nsfw: bool = MISSING,
     ) -> VoiceChannel:
         """|coro|
 
@@ -1280,6 +1645,17 @@ class Guild(Hashable):
         reason: Optional[:class:`str`]
             The reason for creating this channel. Shows up on the audit log.
 
+        slowmode_delay: :class:`int`
+            Specifies the slowmode rate limit for user in this channel, in seconds.
+            A value of ``0`` disables slowmode. The maximum value possible is ``21600``.
+
+            .. versionadded:: 2.7
+
+        nsfw: :class:`bool`
+            Whether the channel is marked as NSFW.
+
+            .. versionadded:: 2.7
+
         Returns
         -------
         :class:`VoiceChannel`
@@ -1310,6 +1686,12 @@ class Guild(Hashable):
         if video_quality_mode is not MISSING:
             options["video_quality_mode"] = video_quality_mode.value
 
+        if slowmode_delay is not MISSING:
+            options["rate_limit_per_user"] = slowmode_delay
+
+        if nsfw is not MISSING:
+            options["nsfw"] = nsfw
+
         data = await self._create_channel(
             name,
             overwrites=overwrites,
@@ -1333,6 +1715,12 @@ class Guild(Hashable):
         overwrites: dict[Role | Member, PermissionOverwrite] = MISSING,
         category: CategoryChannel | None = None,
         reason: str | None = None,
+        bitrate: int = MISSING,
+        user_limit: int = MISSING,
+        rtc_region: VoiceRegion | None = MISSING,
+        video_quality_mode: VideoQualityMode = MISSING,
+        slowmode_delay: int = MISSING,
+        nsfw: bool = MISSING,
     ) -> StageChannel:
         """|coro|
 
@@ -1358,6 +1746,38 @@ class Guild(Hashable):
         reason: Optional[:class:`str`]
             The reason for creating this channel. Shows up on the audit log.
 
+        bitrate: :class:`int`
+            The channel's preferred audio bitrate in bits per second.
+
+            .. versionadded:: 2.7
+
+        user_limit: :class:`int`
+            The channel's limit for number of members that can be in a voice channel.
+
+            .. versionadded:: 2.7
+
+        rtc_region: Optional[:class:`VoiceRegion`]
+            The region for the voice channel's voice communication.
+            A value of ``None`` indicates automatic voice region detection.
+
+            .. versionadded:: 2.7
+
+        video_quality_mode: :class:`VideoQualityMode`
+            The camera video quality for the voice channel's participants.
+
+            .. versionadded:: 2.7
+
+        slowmode_delay: :class:`int`
+            Specifies the slowmode rate limit for user in this channel, in seconds.
+            A value of ``0`` disables slowmode. The maximum value possible is ``21600``.
+
+            .. versionadded:: 2.7
+
+        nsfw: :class:`bool`
+            Whether the channel is marked as NSFW.
+
+            .. versionadded:: 2.7
+
         Returns
         -------
         :class:`StageChannel`
@@ -1378,6 +1798,24 @@ class Guild(Hashable):
         }
         if position is not MISSING:
             options["position"] = position
+
+        if bitrate is not MISSING:
+            options["bitrate"] = bitrate
+
+        if user_limit is not MISSING:
+            options["user_limit"] = user_limit
+
+        if rtc_region is not MISSING:
+            options["rtc_region"] = None if rtc_region is None else str(rtc_region)
+
+        if video_quality_mode is not MISSING:
+            options["video_quality_mode"] = video_quality_mode.value
+
+        if slowmode_delay is not MISSING:
+            options["rate_limit_per_user"] = slowmode_delay
+
+        if nsfw is not MISSING:
+            options["nsfw"] = nsfw
 
         data = await self._create_channel(
             name,
@@ -1405,6 +1843,10 @@ class Guild(Hashable):
         nsfw: bool = MISSING,
         overwrites: dict[Role | Member, PermissionOverwrite] = MISSING,
         default_reaction_emoji: GuildEmoji | int | str = MISSING,
+        available_tags: list[ForumTag] = MISSING,
+        default_sort_order: SortOrder | None = MISSING,
+        default_thread_slowmode_delay: int | None = MISSING,
+        default_auto_archive_duration: int = MISSING,
     ) -> ForumChannel:
         """|coro|
 
@@ -1441,9 +1883,9 @@ class Guild(Hashable):
             The new channel's topic.
         slowmode_delay: :class:`int`
             Specifies the slowmode rate limit for user in this channel, in seconds.
-            The maximum value possible is `21600`.
+            A value of ``0`` disables slowmode. The maximum value possible is ``21600``.
         nsfw: :class:`bool`
-            To mark the channel as NSFW or not.
+            Whether the channel is marked as NSFW.
         reason: Optional[:class:`str`]
             The reason for creating this channel. Shows up on the audit log.
         default_reaction_emoji: Optional[:class:`GuildEmoji` | :class:`int` | :class:`str`]
@@ -1452,6 +1894,26 @@ class Guild(Hashable):
             :class:`GuildEmoji`, snowflake ID, string representation (eg. '<a:emoji_name:emoji_id>').
 
             .. versionadded:: v2.5
+
+        available_tags: List[:class:`ForumTag`]
+            The set of tags that can be used in a forum channel.
+
+            .. versionadded:: 2.7
+
+        default_sort_order: Optional[:class:`SortOrder`]
+            The default sort order type used to order posts in this channel.
+
+            .. versionadded:: 2.7
+
+        default_thread_slowmode_delay: Optional[:class:`int`]
+            The initial slowmode delay to set on newly created threads in this channel.
+
+            .. versionadded:: 2.7
+
+        default_auto_archive_duration: :class:`int`
+            The default auto archive duration in minutes for threads created in this channel.
+
+            .. versionadded:: 2.7
 
         Returns
         -------
@@ -1501,6 +1963,20 @@ class Guild(Hashable):
         if nsfw is not MISSING:
             options["nsfw"] = nsfw
 
+        if available_tags is not MISSING:
+            options["available_tags"] = [tag.to_dict() for tag in available_tags]
+
+        if default_sort_order is not MISSING:
+            options["default_sort_order"] = (
+                default_sort_order.value if default_sort_order else None
+            )
+
+        if default_thread_slowmode_delay is not MISSING:
+            options["default_thread_slowmode_delay"] = default_thread_slowmode_delay
+
+        if default_auto_archive_duration is not MISSING:
+            options["default_auto_archive_duration"] = default_auto_archive_duration
+
         if default_reaction_emoji is not MISSING:
             if isinstance(
                 default_reaction_emoji, _EmojiTag
@@ -1512,13 +1988,17 @@ class Guild(Hashable):
                 )
             elif isinstance(default_reaction_emoji, str):
                 default_reaction_emoji = PartialEmoji.from_str(default_reaction_emoji)
+            elif default_reaction_emoji is None:
+                pass
             else:
                 raise InvalidArgument(
-                    "default_reaction_emoji must be of type: GuildEmoji | int | str"
+                    "default_reaction_emoji must be of type: GuildEmoji | int | str | None"
                 )
 
             options["default_reaction_emoji"] = (
                 default_reaction_emoji._to_forum_reaction_payload()
+                if default_reaction_emoji
+                else None
             )
 
         data = await self._create_channel(
@@ -1602,43 +2082,6 @@ class Guild(Hashable):
         """
         await self._state.http.leave_guild(self.id)
 
-    async def delete(self) -> None:
-        """|coro|
-
-        Deletes the guild. You must be the guild owner to delete the
-        guild.
-
-        Raises
-        ------
-        HTTPException
-            Deleting the guild failed.
-        Forbidden
-            You do not have permissions to delete the guild.
-        """
-        await self._state.http.delete_guild(self.id)
-
-    async def set_mfa_required(self, required: bool, *, reason: str = None) -> None:
-        """|coro|
-
-        Set whether it is required to have MFA enabled on your account
-        to perform moderation actions. You must be the guild owner to do this.
-
-        Parameters
-        ----------
-        required: :class:`bool`
-            Whether MFA should be required to perform moderation actions.
-        reason: :class:`str`
-            The reason to show up in the audit log.
-
-        Raises
-        ------
-        HTTPException
-            The operation failed.
-        Forbidden
-            You are not the owner of the guild.
-        """
-        await self._state.http.edit_guild_mfa(self.id, required, reason=reason)
-
     async def edit(
         self,
         *,
@@ -1651,7 +2094,6 @@ class Guild(Hashable):
         discovery_splash: bytes | None = MISSING,
         community: bool = MISSING,
         afk_channel: VoiceChannel | None = MISSING,
-        owner: Snowflake = MISSING,
         afk_timeout: int = MISSING,
         default_notifications: NotificationLevel = MISSING,
         verification_level: VerificationLevel = MISSING,
@@ -1663,6 +2105,9 @@ class Guild(Hashable):
         public_updates_channel: TextChannel | None = MISSING,
         premium_progress_bar_enabled: bool = MISSING,
         disable_invites: bool = MISSING,
+        discoverable: bool = MISSING,
+        disable_raid_alerts: bool = MISSING,
+        enable_activity_feed: bool = MISSING,
     ) -> Guild:
         r"""|coro|
 
@@ -1712,9 +2157,6 @@ class Guild(Hashable):
             The new channel that is the AFK channel. Could be ``None`` for no AFK channel.
         afk_timeout: :class:`int`
             The number of seconds until someone is moved to the AFK channel.
-        owner: :class:`Member`
-            The new owner of the guild to transfer ownership to. Note that you must
-            be owner of the guild to do this.
         verification_level: :class:`VerificationLevel`
             The new verification level for the guild.
         default_notifications: :class:`NotificationLevel`
@@ -1740,6 +2182,12 @@ class Guild(Hashable):
             Whether the guild should have premium progress bar enabled.
         disable_invites: :class:`bool`
             Whether the guild should have server invites enabled or disabled.
+        discoverable: :class:`bool`
+            Whether the guild should be discoverable in the discover tab.
+        disable_raid_alerts: :class:`bool`
+            Whether activity alerts for the guild should be disabled.
+        enable_activity_feed: class:`bool`
+            Whether the guild's user activity feed should be enabled.
         reason: Optional[:class:`str`]
             The reason for editing this guild. Shows up on the audit log.
 
@@ -1751,8 +2199,7 @@ class Guild(Hashable):
             Editing the guild failed.
         InvalidArgument
             The image format passed in to ``icon`` is invalid. It must be
-            PNG or JPG. This is also raised if you are not the owner of the
-            guild and request an ownership transfer.
+            PNG or JPG.
 
         Returns
         --------
@@ -1829,14 +2276,6 @@ class Guild(Hashable):
             else:
                 fields["public_updates_channel_id"] = public_updates_channel.id
 
-        if owner is not MISSING:
-            if self.owner_id != self._state.self_id:
-                raise InvalidArgument(
-                    "To transfer ownership you must be the owner of the guild."
-                )
-
-            fields["owner_id"] = owner.id
-
         if verification_level is not MISSING:
             if not isinstance(verification_level, VerificationLevel):
                 raise InvalidArgument(
@@ -1861,8 +2300,12 @@ class Guild(Hashable):
 
             fields["system_channel_flags"] = system_channel_flags.value
 
+        if premium_progress_bar_enabled is not MISSING:
+            fields["premium_progress_bar_enabled"] = premium_progress_bar_enabled
+
+        features: list[GuildFeature] = self.features.copy()
+
         if community is not MISSING:
-            features = self.features.copy()
             if community:
                 if (
                     "rules_channel_id" in fields
@@ -1872,8 +2315,7 @@ class Guild(Hashable):
                         features.append("COMMUNITY")
                 else:
                     raise InvalidArgument(
-                        "community field requires both rules_channel and"
-                        " public_updates_channel fields to be provided"
+                        "community field requires both rules_channel and public_updates_channel fields to be provided"
                     )
             else:
                 if "COMMUNITY" in features:
@@ -1883,20 +2325,43 @@ class Guild(Hashable):
                         fields["public_updates_channel_id"] = None
                     features.remove("COMMUNITY")
 
-            fields["features"] = features
-
-        if premium_progress_bar_enabled is not MISSING:
-            fields["premium_progress_bar_enabled"] = premium_progress_bar_enabled
-
         if disable_invites is not MISSING:
-            features = self.features.copy()
             if disable_invites:
-                if not "INVITES_DISABLED" in features:
+                if "INVITES_DISABLED" not in features:
                     features.append("INVITES_DISABLED")
             else:
                 if "INVITES_DISABLED" in features:
                     features.remove("INVITES_DISABLED")
 
+        if discoverable is not MISSING:
+            if discoverable:
+                if "DISCOVERABLE" not in features:
+                    features.append("DISCOVERABLE")
+            else:
+                if "DISCOVERABLE" in features:
+                    features.remove("DISCOVERABLE")
+
+        if disable_raid_alerts is not MISSING:
+            if disable_raid_alerts:
+                if "RAID_ALERTS_DISABLED" not in features:
+                    features.append("RAID_ALERTS_DISABLED")
+            else:
+                if "RAID_ALERTS_DISABLED" in features:
+                    features.remove("RAID_ALERTS_DISABLED")
+
+        if enable_activity_feed is not MISSING:
+            if enable_activity_feed:
+                if "ACTIVITY_FEED_ENABLED_BY_USER" not in features:
+                    features.append("ACTIVITY_FEED_ENABLED_BY_USER")
+                if "ACTIVITY_FEED_DISABLED_BY_USER" in features:
+                    features.remove("ACTIVITY_FEED_DISABLED_BY_USER")
+            else:
+                if "ACTIVITY_FEED_ENABLED_BY_USER" in features:
+                    features.remove("ACTIVITY_FEED_ENABLED_BY_USER")
+                if "ACTIVITY_FEED_DISABLED_BY_USER" not in features:
+                    features.append("ACTIVITY_FEED_DISABLED_BY_USER")
+
+        if self.features != features:
             fields["features"] = features
 
         data = await http.edit_guild(self.id, reason=reason, **fields)
@@ -2664,6 +3129,26 @@ class Guild(Hashable):
         """
         await self._state.http.delete_guild_sticker(self.id, sticker.id, reason)
 
+    def get_emoji(self, emoji_id: int, /) -> GuildEmoji | None:
+        """Returns an emoji with the given ID.
+
+        .. versionadded:: 2.7
+
+        Parameters
+        ----------
+        emoji_id: int
+            The ID to get.
+
+        Returns
+        -------
+        Optional[:class:`Emoji`]
+            The returned Emoji or ``None`` if not found.
+        """
+        emoji = self._state.get_emoji(emoji_id)
+        if emoji and emoji.guild == self:
+            return emoji
+        return None
+
     async def fetch_emojis(self) -> list[GuildEmoji]:
         r"""|coro|
 
@@ -2881,6 +3366,8 @@ class Guild(Hashable):
         name: str = ...,
         permissions: Permissions = ...,
         colour: Colour | int = ...,
+        colours: RoleColours = ...,
+        holographic: bool = ...,
         hoist: bool = ...,
         mentionable: bool = ...,
         icon: bytes | None = MISSING,
@@ -2895,6 +3382,8 @@ class Guild(Hashable):
         name: str = ...,
         permissions: Permissions = ...,
         color: Colour | int = ...,
+        colors: RoleColours = ...,
+        holographic: bool = ...,
         hoist: bool = ...,
         mentionable: bool = ...,
         icon: bytes | None = ...,
@@ -2908,6 +3397,9 @@ class Guild(Hashable):
         permissions: Permissions = MISSING,
         color: Colour | int = MISSING,
         colour: Colour | int = MISSING,
+        colors: RoleColours = MISSING,
+        colours: RoleColours = MISSING,
+        holographic: bool = MISSING,
         hoist: bool = MISSING,
         mentionable: bool = MISSING,
         reason: str | None = None,
@@ -2971,11 +3463,30 @@ class Guild(Hashable):
         else:
             fields["permissions"] = "0"
 
-        actual_colour = colour or color or Colour.default()
+        actual_colour = colour if colour not in (MISSING, None) else color
+
         if isinstance(actual_colour, int):
-            fields["color"] = actual_colour
+            actual_colour = Colour(actual_colour)
+
+        if actual_colour not in (MISSING, None):
+            utils.warn_deprecated("colour", "colours", "2.7")
+            actual_colours = RoleColours(primary=actual_colour)
+        elif holographic:
+            actual_colours = RoleColours.holographic()
         else:
-            fields["color"] = actual_colour.value
+            actual_colours = colours or colors or RoleColours.default()
+
+        if isinstance(actual_colours, RoleColours):
+            if "ENHANCED_ROLE_COLORS" not in self.features:
+                actual_colours.secondary = None
+                actual_colours.tertiary = None
+            fields["colors"] = actual_colours._to_dict()
+        else:
+            raise InvalidArgument(
+                "colours parameter must be of type RoleColours, not {0.__class__.__name__}".format(
+                    actual_colours
+                )
+            )
 
         if hoist is not MISSING:
             fields["hoist"] = hoist
@@ -3770,8 +4281,8 @@ class Guild(Hashable):
         *,
         name: str,
         description: str = MISSING,
-        start_time: datetime,
-        end_time: datetime = MISSING,
+        start_time: datetime.datetime,
+        end_time: datetime.datetime = MISSING,
         location: str | int | VoiceChannel | StageChannel | ScheduledEventLocation,
         privacy_level: ScheduledEventPrivacyLevel = ScheduledEventPrivacyLevel.guild_only,
         reason: str | None = None,
@@ -4045,6 +4556,52 @@ class Guild(Hashable):
         new = await self._state.http.edit_onboarding(self.id, fields, reason=reason)
         return Onboarding(data=new, guild=self)
 
+    async def modify_incident_actions(
+        self,
+        *,
+        invites_disabled_until: datetime.datetime | None = MISSING,
+        dms_disabled_until: datetime.datetime | None = MISSING,
+        reason: str | None = MISSING,
+    ) -> IncidentsData:
+        """|coro|
+
+        Modify the guild's incident actions, controlling when invites or DMs
+        are re-enabled after being temporarily disabled. Requires
+        the :attr:`~Permissions.manage_guild` permission.
+
+        Parameters
+        ----------
+        invites_disabled_until: Optional[:class:`datetime.datetime`]
+            The ISO8601 timestamp indicating when invites will be enabled again,
+            or ``None`` to enable invites immediately.
+        dms_disabled_until: Optional[:class:`datetime.datetime`]
+            The ISO8601 timestamp indicating when DMs will be enabled again,
+            or ``None`` to enable DMs immediately.
+        reason: Optional[:class:`str`]
+            The reason for this action, used for the audit log.
+
+        Returns
+        -------
+        :class:`IncidentsData`
+            The updated incidents data for the guild.
+        """
+
+        fields: ModifyIncidentsPayload = {}
+        if invites_disabled_until is not MISSING:
+            fields["invites_disabled_until"] = (
+                invites_disabled_until and invites_disabled_until.isoformat()
+            )
+
+        if dms_disabled_until is not MISSING:
+            fields["dms_disabled_until"] = (
+                dms_disabled_until and dms_disabled_until.isoformat()
+            )
+
+        new = await self._state.http.modify_guild_incident_actions(
+            self.id, fields, reason=reason
+        )
+        return IncidentsData(data=new)
+
     async def delete_auto_moderation_rule(
         self,
         id: int,
@@ -4149,3 +4706,20 @@ class Guild(Hashable):
             guild_id=self.id,
             exclude_ended=exclude_ended,
         )
+
+    def get_sound(self, sound_id: int) -> SoundboardSound | None:
+        """Returns a sound with the given ID.
+
+        .. versionadded :: 2.7
+
+        Parameters
+        ----------
+        sound_id: :class:`int`
+            The ID to search for.
+
+        Returns
+        -------
+        Optional[:class:`SoundboardSound`]
+            The sound or ``None`` if not found.
+        """
+        return self._sounds.get(sound_id)
