@@ -33,6 +33,8 @@ from .mixins import Hashable
 if TYPE_CHECKING:
     import datetime
 
+    from .abc import Snowflake
+
     SupportsIntCast = Union[SupportsInt, str, bytes, bytearray]
 
 __all__ = ("Object",)
@@ -70,9 +72,11 @@ class Object(Hashable):
     ----------
     id: :class:`int`
         The ID of the object.
+    type: type[:class:`abc.Snowflake`]
+        The model this object's ID is based off.
     """
 
-    def __init__(self, id: SupportsIntCast):
+    def __init__(self, id: SupportsIntCast, type: type[Snowflake] = utils.MISSING):
         try:
             id = int(id)
         except ValueError:
@@ -81,6 +85,7 @@ class Object(Hashable):
             ) from None
         else:
             self.id = id
+        self.type: type[Snowflake] = type or self.__class__
 
     def __repr__(self) -> str:
         return f"<Object id={self.id!r}>"
