@@ -24,11 +24,12 @@ DEALINGS IN THE SOFTWARE.
 
 from __future__ import annotations
 
+from functools import cached_property
 from typing import TYPE_CHECKING, Any
 
 from .enums import OnboardingMode, PromptType, try_enum
 from .partial_emoji import PartialEmoji
-from .utils import MISSING, cached_property, generate_snowflake, get
+from .utils import MISSING, generate_snowflake, get
 
 if TYPE_CHECKING:
     from .abc import Snowflake
