@@ -28,7 +28,13 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Callable, Iterable
 
 from .abc import Messageable, _purge_messages_helper
-from .enums import ChannelType, try_enum
+from .enums import (
+    ChannelType,
+)
+from .enums import ThreadArchiveDuration as ThreadArchiveDurationEnum
+from .enums import (
+    try_enum,
+)
 from .errors import ClientException
 from .flags import ChannelFlags
 from .mixins import Hashable
@@ -602,7 +608,9 @@ class Thread(Messageable, Hashable):
         locked: bool = MISSING,
         invitable: bool = MISSING,
         slowmode_delay: int = MISSING,
-        auto_archive_duration: ThreadArchiveDuration = MISSING,
+        auto_archive_duration: (
+            ThreadArchiveDuration | ThreadArchiveDurationEnum
+        ) = MISSING,
         pinned: bool = MISSING,
         applied_tags: list[ForumTag] = MISSING,
         reason: str | None = None,
@@ -632,6 +640,7 @@ class Thread(Messageable, Hashable):
         auto_archive_duration: :class:`int`
             The new duration in minutes before a thread is automatically archived for inactivity.
             Must be one of ``60``, ``1440``, ``4320``, or ``10080``.
+            :class:`ThreadArchiveDuration` can be used alternatively.
         slowmode_delay: :class:`int`
             Specifies the slowmode rate limit for user in this thread, in seconds.
             A value of ``0`` disables slowmode. The maximum value possible is ``21600``.
