@@ -81,6 +81,7 @@ class Attachment(TypedDict):
     waveform: NotRequired[str]
     flags: NotRequired[int]
     title: NotRequired[str]
+    ephemeral: NotRequired[bool]
 
 
 MessageActivityType = Literal[1, 2, 3, 5]
@@ -99,7 +100,11 @@ class MessageApplication(TypedDict):
     name: str
 
 
+MessageReferenceType = Literal[0, 1]
+
+
 class MessageReference(TypedDict, total=False):
+    type: NotRequired[MessageReferenceType]
     message_id: Snowflake
     channel_id: Snowflake
     guild_id: Snowflake
@@ -109,6 +114,29 @@ class MessageReference(TypedDict, total=False):
 MessageType = Literal[
     0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 14, 15, 18, 19, 20, 21, 22, 23, 24
 ]
+
+
+class MessageCall(TypedDict):
+    participants: SnowflakeList
+    ended_timestamp: NotRequired[str]
+
+
+class ForwardedMessage(TypedDict):
+    type: MessageType
+    content: str
+    embeds: list[Embed]
+    attachments: list[Attachment]
+    timestamp: str
+    edited_timestamp: str | None
+    flags: NotRequired[int]
+    mentions: list[UserWithMember]
+    mention_roles: SnowflakeList
+    sticker_items: NotRequired[list[StickerItem]]
+    components: NotRequired[list[Component]]
+
+
+class MessageSnapshot(TypedDict):
+    message: ForwardedMessage
 
 
 class Message(TypedDict):
@@ -144,6 +172,18 @@ class Message(TypedDict):
     pinned: bool
     type: MessageType
     poll: Poll
+    call: MessageCall
+    message_snapshots: NotRequired[list[MessageSnapshot]]
+
+
+class MessagePin(TypedDict):
+    pinned_at: str
+    message: Message
+
+
+class MessagePinPagination(TypedDict):
+    items: list[MessagePin]
+    has_more: bool
 
 
 AllowedMentionType = Literal["roles", "users", "everyone"]
@@ -154,8 +194,3 @@ class AllowedMentions(TypedDict):
     roles: SnowflakeList
     users: SnowflakeList
     replied_user: bool
-
-
-class MessageCall(TypedDict):
-    participants: SnowflakeList
-    ended_timestamp: NotRequired[str]
