@@ -32,7 +32,7 @@ from typing import TYPE_CHECKING, Callable, TypeVar
 from ..components import Button as ButtonComponent
 from ..enums import ButtonStyle, ComponentType
 from ..partial_emoji import PartialEmoji, _EmojiTag
-from .item import Item, ItemCallbackType
+from .item import ItemCallbackType, ViewItem
 
 __all__ = (
     "Button",
@@ -41,13 +41,14 @@ __all__ = (
 
 if TYPE_CHECKING:
     from ..emoji import AppEmoji, GuildEmoji
-    from .view import View
+    from ..types.components import ButtonComponent as ButtonComponentPayload
+    from .view import BaseView
 
 B = TypeVar("B", bound="Button")
-V = TypeVar("V", bound="View", covariant=True)
+V = TypeVar("V", bound="BaseView", covariant=True)
 
 
-class Button(Item[V]):
+class Button(ViewItem[V]):
     """Represents a UI button.
 
     .. versionadded:: 2.0
@@ -75,6 +76,13 @@ class Button(Item[V]):
         like to control the relative positioning of the row then passing an index is advised.
         For example, row=1 will show up before row=2. Defaults to ``None``, which is automatic
         ordering. The row number must be between 0 and 4 (i.e. zero indexed).
+
+        .. warning::
+
+            This parameter does not work in :class:`ActionRow`.
+
+    id: Optional[:class:`int`]
+        The button's ID.
     """
 
     __item_repr_attributes__: tuple[str, ...] = (
@@ -85,6 +93,8 @@ class Button(Item[V]):
         "emoji",
         "sku_id",
         "row",
+        "custom_id",
+        "id",
     )
 
     def __init__(
@@ -98,7 +108,10 @@ class Button(Item[V]):
         emoji: str | GuildEmoji | AppEmoji | PartialEmoji | None = None,
         sku_id: int | None = None,
         row: int | None = None,
+        id: int | None = None,
     ):
+        self._row: int | None = None
+        self._rendered_row: int | None = None
         super().__init__()
         if label and len(str(label)) > 80:
             raise ValueError("label must be 80 characters or fewer")
@@ -136,8 +149,8 @@ class Button(Item[V]):
                     f" {emoji.__class__}"
                 )
 
-        self._underlying = ButtonComponent._raw_construct(
-            type=ComponentType.button,
+        self.row = row
+        self._underlying = self._generate_underlying(
             custom_id=custom_id,
             url=url,
             disabled=disabled,
@@ -145,17 +158,41 @@ class Button(Item[V]):
             style=style,
             emoji=emoji,
             sku_id=sku_id,
+            id=id,
         )
-        self.row = row
+
+    def _generate_underlying(
+        self,
+        style: ButtonStyle | None = None,
+        label: str | None = None,
+        disabled: bool = False,
+        custom_id: str | None = None,
+        url: str | None = None,
+        emoji: str | GuildEmoji | AppEmoji | PartialEmoji | None = None,
+        sku_id: int | None = None,
+        id: int | None = None,
+    ) -> ButtonComponent:
+        super()._generate_underlying(ButtonComponent)
+        return ButtonComponent._raw_construct(
+            type=ComponentType.button,
+            custom_id=custom_id or self.custom_id,
+            url=url or self.url,
+            disabled=disabled or self.disabled,
+            label=label or self.label,
+            style=style or self.style,
+            emoji=emoji or self.emoji,
+            sku_id=sku_id or self.sku_id,
+            id=id or self.id,
+        )
 
     @property
     def style(self) -> ButtonStyle:
         """The style of the button."""
-        return self._underlying.style
+        return self.underlying.style
 
     @style.setter
     def style(self, value: ButtonStyle):
-        self._underlying.style = value
+        self.underlying.style = value
 
     @property
     def custom_id(self) -> str | None:
@@ -163,7 +200,7 @@ class Button(Item[V]):
 
         If this button is for a URL, it does not have a custom ID.
         """
-        return self._underlying.custom_id
+        return self.underlying.custom_id
 
     @custom_id.setter
     def custom_id(self, value: str | None):
@@ -171,52 +208,53 @@ class Button(Item[V]):
             raise TypeError("custom_id must be None or str")
         if value and len(value) > 100:
             raise ValueError("custom_id must be 100 characters or fewer")
-        self._underlying.custom_id = value
+        self.underlying.custom_id = value
+        self._provided_custom_id = value is not None
 
     @property
     def url(self) -> str | None:
         """The URL this button sends you to."""
-        return self._underlying.url
+        return self.underlying.url
 
     @url.setter
     def url(self, value: str | None):
         if value is not None and not isinstance(value, str):
             raise TypeError("url must be None or str")
-        self._underlying.url = value
+        self.underlying.url = value
 
     @property
     def disabled(self) -> bool:
         """Whether the button is disabled or not."""
-        return self._underlying.disabled
+        return self.underlying.disabled
 
     @disabled.setter
     def disabled(self, value: bool):
-        self._underlying.disabled = bool(value)
+        self.underlying.disabled = bool(value)
 
     @property
     def label(self) -> str | None:
         """The label of the button, if available."""
-        return self._underlying.label
+        return self.underlying.label
 
     @label.setter
     def label(self, value: str | None):
         if value and len(str(value)) > 80:
             raise ValueError("label must be 80 characters or fewer")
-        self._underlying.label = str(value) if value is not None else value
+        self.underlying.label = str(value) if value is not None else value
 
     @property
     def emoji(self) -> PartialEmoji | None:
         """The emoji of the button, if available."""
-        return self._underlying.emoji
+        return self.underlying.emoji
 
     @emoji.setter
     def emoji(self, value: str | GuildEmoji | AppEmoji | PartialEmoji | None):  # type: ignore
         if value is None:
-            self._underlying.emoji = None
+            self.underlying.emoji = None
         elif isinstance(value, str):
-            self._underlying.emoji = PartialEmoji.from_str(value)
+            self.underlying.emoji = PartialEmoji.from_str(value)
         elif isinstance(value, _EmojiTag):
-            self._underlying.emoji = value._to_partial()
+            self.underlying.emoji = value._to_partial()
         else:
             raise TypeError(
                 "expected str, GuildEmoji, AppEmoji, or PartialEmoji, received"
@@ -226,16 +264,57 @@ class Button(Item[V]):
     @property
     def sku_id(self) -> int | None:
         """The ID of the SKU this button refers to."""
-        return self._underlying.sku_id
+        return self.underlying.sku_id
 
     @sku_id.setter
     def sku_id(self, value: int | None):  # type: ignore
         if value is None:
-            self._underlying.sku_id = None
+            self.underlying.sku_id = None
         elif isinstance(value, int):
-            self._underlying.sku_id = value
+            self.underlying.sku_id = value
         else:
             raise TypeError(f"expected int or None, received {value.__class__} instead")
+
+    @property
+    def width(self) -> int:
+        """Gets the width of the item in the UI layout.
+
+        The width determines how much horizontal space this item occupies within its row.
+
+        Returns
+        -------
+        :class:`int`
+            The width of the item. Buttons have a width of 1.
+        """
+        return 1
+
+    @property
+    def row(self) -> int | None:
+        """Gets or sets the row position of this item within its parent view.
+
+        The row position determines the vertical placement of the item in the UI.
+        The value must be an integer between 0 and 4 (inclusive), or ``None`` to indicate
+        that no specific row is set.
+        This attribute is not compatible with :class:`discord.ui.DesignerView`.
+
+        Returns
+        -------
+        Optional[:class:`int`]
+            The row position of the item, or ``None`` if not explicitly set.
+
+        Raises
+        ------
+        ValueError
+            If the row value is not ``None`` and is outside the range [0, 4].
+        """
+        return self._row
+
+    @row.setter
+    def row(self, value: int | None):
+        if value is None or 5 > value >= 0:
+            self._row = value
+        else:
+            raise ValueError("row cannot be negative or greater than or equal to 5")
 
     @classmethod
     def from_component(cls: type[B], button: ButtonComponent) -> B:
@@ -248,17 +327,19 @@ class Button(Item[V]):
             emoji=button.emoji,
             sku_id=button.sku_id,
             row=None,
+            id=button.id,
         )
 
-    @property
-    def type(self) -> ComponentType:
-        return self._underlying.type
-
-    def to_component_dict(self):
-        return self._underlying.to_dict()
+    def to_component_dict(self) -> ButtonComponentPayload:
+        return super().to_component_dict()
 
     def is_dispatchable(self) -> bool:
-        return self.custom_id is not None
+        return (self.custom_id is not None) and (
+            bool(self.view._store) if self.view else True
+        )
+
+    def is_storable(self) -> bool:
+        return self.is_dispatchable()
 
     def is_persistent(self) -> bool:
         if self.style is ButtonStyle.link:
@@ -266,7 +347,7 @@ class Button(Item[V]):
         return super().is_persistent()
 
     def refresh_component(self, button: ButtonComponent) -> None:
-        self._underlying = button
+        self.underlying = button
 
 
 def button(
@@ -277,11 +358,12 @@ def button(
     style: ButtonStyle = ButtonStyle.secondary,
     emoji: str | GuildEmoji | AppEmoji | PartialEmoji | None = None,
     row: int | None = None,
-) -> Callable[[ItemCallbackType], ItemCallbackType]:
+    id: int | None = None,
+) -> Callable[[ItemCallbackType[Button[V]]], Button[V]]:
     """A decorator that attaches a button to a component.
 
     The function being decorated should have three parameters, ``self`` representing
-    the :class:`discord.ui.View`, the :class:`discord.ui.Button` being pressed and
+    the :class:`discord.ui.View`, :class:`discord.ui.ActionRow` or :class:`discord.ui.Section`, the :class:`discord.ui.Button` being pressed, and
     the :class:`discord.Interaction` you receive.
 
     .. note::
@@ -311,6 +393,10 @@ def button(
         like to control the relative positioning of the row then passing an index is advised.
         For example, row=1 will show up before row=2. Defaults to ``None``, which is automatic
         ordering. The row number must be between 0 and 4 (i.e. zero indexed).
+
+        .. warning::
+
+            This parameter does not work in :class:`ActionRow`.
     """
 
     def decorator(func: ItemCallbackType) -> ItemCallbackType:
@@ -326,7 +412,8 @@ def button(
             "label": label,
             "emoji": emoji,
             "row": row,
+            "id": id,
         }
         return func
 
-    return decorator
+    return decorator  # type: ignore # lie to the type checkers, because after a View is instated, the button callback is converted into a Button instance
