@@ -29,48 +29,51 @@ from typing import TYPE_CHECKING, Any, Callable, Coroutine, Generic, TypeVar
 
 from ..interactions import Interaction
 
-__all__ = ("Item",)
+__all__ = (
+    "Item",
+    "ViewItem",
+    "ModalItem",
+)
+
+from ..utils import warn_deprecated
 
 if TYPE_CHECKING:
     from ..components import Component
     from ..enums import ComponentType
-    from .view import View
+    from .core import ItemInterface
+    from .modal import BaseModal
+    from .view import BaseView
 
 I = TypeVar("I", bound="Item")
-V = TypeVar("V", bound="View", covariant=True)
+T = TypeVar("T", bound="ItemInterface", covariant=True)
+V = TypeVar("V", bound="BaseView", covariant=True)
+M = TypeVar("M", bound="BaseModal", covariant=True)
 ItemCallbackType = Callable[[Any, I, Interaction], Coroutine[Any, Any, Any]]
 
 
-class Item(Generic[V]):
+class Item(Generic[T]):
     """Represents the base UI item that all UI components inherit from.
 
-    The current UI items supported are:
-
-    - :class:`discord.ui.Button`
-    - :class:`discord.ui.Select`
-
     .. versionadded:: 2.0
+
+    .. versionchanged:: 2.7
+        Now used as base class for :class:`ViewItem` and :class:`ModalItem`.
     """
 
-    __item_repr_attributes__: tuple[str, ...] = ("row",)
+    __item_repr_attributes__: tuple[str, ...] = ("id",)
 
     def __init__(self):
-        self._view: V | None = None
-        self._row: int | None = None
-        self._rendered_row: int | None = None
-        # This works mostly well but there is a gotcha with
-        # the interaction with from_component, since that technically provides
-        # a custom_id most dispatchable items would get this set to True even though
-        # it might not be provided by the library user. However, this edge case doesn't
-        # actually affect the intended purpose of this check because from_component is
-        # only called upon edit and we're mainly interested during initial creation time.
+        self._underlying: Component | None = None
         self._provided_custom_id: bool = False
+        self.parent: Item | ItemInterface | None = None
 
     def to_component_dict(self) -> dict[str, Any]:
-        raise NotImplementedError
+        if not self.underlying:
+            raise NotImplementedError
+        return self.underlying.to_dict()
 
     def refresh_component(self, component: Component) -> None:
-        return None
+        self.underlying = component
 
     def refresh_state(self, interaction: Interaction) -> None:
         return None
@@ -80,14 +83,35 @@ class Item(Generic[V]):
         return cls()
 
     @property
+    def underlying(self) -> Component:
+        return self._underlying
+
+    @underlying.setter
+    def underlying(self, value: Component) -> None:
+        self._underlying = value
+
+    @property
     def type(self) -> ComponentType:
-        raise NotImplementedError
+        if not self.underlying:
+            raise NotImplementedError
+        return self.underlying.type
+
+    def _generate_underlying(self, cls: type[Component]) -> Component:
+        if not self._underlying:
+            self._underlying = cls._raw_construct()
+        return self._underlying
 
     def is_dispatchable(self) -> bool:
         return False
 
+    def is_storable(self) -> bool:
+        return False
+
     def is_persistent(self) -> bool:
-        return self._provided_custom_id
+        return not self.is_dispatchable() or self._provided_custom_id
+
+    def copy_text(self) -> str:
+        return ""
 
     def __repr__(self) -> str:
         attrs = " ".join(
@@ -96,60 +120,92 @@ class Item(Generic[V]):
         return f"<{self.__class__.__name__} {attrs}>"
 
     @property
-    def row(self) -> int | None:
-        """Gets or sets the row position of this item within its parent view.
+    def id(self) -> int | None:
+        """Gets this item's ID.
 
-        The row position determines the vertical placement of the item in the UI.
-        The value must be an integer between 0 and 4 (inclusive), or ``None`` to indicate
-        that no specific row is set.
+        This can be set by the user when constructing an Item. If not, Discord will automatically provide one when the item's parent is sent.
 
         Returns
         -------
         Optional[:class:`int`]
-            The row position of the item, or ``None`` if not explicitly set.
-
-        Raises
-        ------
-        ValueError
-            If the row value is not ``None`` and is outside the range [0, 4].
+            The ID of this item, or ``None`` if the user didn't set one.
         """
-        return self._row
+        return self.underlying and self.underlying.id
+
+    @id.setter
+    def id(self, value) -> None:
+        if not self.underlying:
+            return
+        self.underlying.id = value
+
+
+class ViewItem(Item[V]):
+    """Represents an item used in Views.
+
+    The following are the original items supported in :class:`discord.ui.View`:
+
+    - :class:`discord.ui.Button`
+    - :class:`discord.ui.Select`
+
+    And the following are new items under the "Components V2" specification for use in :class:`discord.ui.DesignerView`:
+
+    - :class:`discord.ui.Section`
+    - :class:`discord.ui.TextDisplay`
+    - :class:`discord.ui.Thumbnail`
+    - :class:`discord.ui.MediaGallery`
+    - :class:`discord.ui.File`
+    - :class:`discord.ui.Separator`
+    - :class:`discord.ui.Container`
+
+    Additionally, :class:`discord.ui.ActionRow` should be used in :class:`discord.ui.DesignerView` to support :class:`discord.ui.Button` and :class:`discord.ui.Select`.
+
+    .. versionadded:: 2.7
+    """
+
+    def __init__(self):
+        super().__init__()
+        self._view: V | None = None
+        self.parent: ViewItem | BaseView | None = None
+
+    @property
+    def row(self) -> int | None:
+        warn_deprecated("Accessing .row from CV2 Items", since="2.7.1", removed="3.0")
+        return None
 
     @row.setter
-    def row(self, value: int | None):
-        if value is None:
-            self._row = None
-        elif 5 > value >= 0:
-            self._row = value
-        else:
-            raise ValueError("row cannot be negative or greater than or equal to 5")
+    def row(self, value: int | None) -> None:
+        warn_deprecated("Setting .row on CV2 Items", since="2.7.1", removed="3.0")
 
     @property
     def width(self) -> int:
-        """Gets the width of the item in the UI layout.
-
-        The width determines how much horizontal space this item occupies within its row.
-
-        Returns
-        -------
-        :class:`int`
-            The width of the item. Defaults to 1.
-        """
         return 1
 
     @property
     def view(self) -> V | None:
         """Gets the parent view associated with this item.
 
-        The view refers to the container that holds this item. This is typically set
+        The view refers to the structure that holds this item. This is typically set
         automatically when the item is added to a view.
 
         Returns
         -------
-        Optional[:class:`View`]
+        Optional[:class:`BaseView`]
             The parent view of this item, or ``None`` if the item is not attached to any view.
         """
-        return self._view
+        if self._view:
+            return self._view
+        if self.parent:
+            from .view import BaseView
+
+            if isinstance(self.parent, BaseView):
+                return self.parent
+            return self.parent.view
+        return None
+
+    @view.setter
+    def view(self, value: V | None) -> None:
+        warn_deprecated("Manually setting .view", since="2.7", removed="3.0")
+        self._view = value
 
     async def callback(self, interaction: Interaction):
         """|coro|
@@ -163,3 +219,46 @@ class Item(Generic[V]):
         interaction: :class:`.Interaction`
             The interaction that triggered this UI item.
         """
+
+
+class ModalItem(Item[M]):
+    """Represents an item used in Modals.
+
+    :class:`discord.ui.InputText` is the original item supported in :class:`discord.ui.Modal`.
+
+    The following are newly available in :class:`discord.ui.DesignerModal`:
+
+    - :class:`discord.ui.Label`
+    - :class:`discord.ui.TextDisplay`
+
+    And :class:`discord.ui.Label` should be used in :class:`discord.ui.DesignerModal` to support the following items:
+    - :class:`discord.ui.InputText`
+    - :class:`discord.ui.Select`
+    - :class:`discord.ui.FileUpload`
+
+    .. versionadded:: 2.7
+    """
+
+    def __init__(self):
+        super().__init__()
+        self._modal: M | None = None
+        self.parent: ModalItem | BaseModal | None = self.modal
+
+    def refresh_from_modal(self, interaction: Interaction, data: dict) -> None:
+        return None
+
+    @property
+    def modal(self) -> M | None:
+        """Gets the parent modal associated with this item. This is typically set
+        automatically when the item is added to a modal.
+
+        Returns
+        -------
+        Optional[:class:`BaseModal`]
+            The parent modal of this item, or ``None`` if the item is not attached to any modal.
+        """
+        return self._modal
+
+    @modal.setter
+    def modal(self, value) -> None:
+        self._modal = value
