@@ -11,11 +11,14 @@ class DropdownView(discord.ui.View):
         placeholder="Select channels...", min_values=1, max_values=3
     )  # Users can select a maximum of 3 channels in the dropdown
     async def channel_select_dropdown(
-        self, select: discord.ui.Select, interaction: discord.Interaction
+        self, select: discord.ui.ChannelSelect, interaction: discord.Interaction
     ) -> None:
+        # update the select default values to the chosen values
+        select.default_values = select.values  # this is a list of GuildChannels
         await interaction.response.send_message(
-            f"You selected the following channels:"
-            + f", ".join(f"{channel.mention}" for channel in select.values)
+            "You selected the following channels:"
+            + ", ".join(f"{channel.mention}" for channel in select.values),
+            view=self,
         )
 
 
